% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{add_message.chat}
\alias{add_message.chat}
\title{Add a message to a chat object}
\usage{
\method{add_message}{chat}(x, content, role = "user")
}
\arguments{
\item{x}{A chat object}

\item{content}{The content of the message}

\item{role}{The role of the message (defaults to "user")}
}
\description{
Add a message to a chat object
}
