% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{add_message.GPT_thread}
\alias{add_message.GPT_thread}
\title{Add a message to a GPT_thread}
\usage{
\method{add_message}{GPT_thread}(x, content, role = "user", files = NULL, metadata = NULL)
}
\arguments{
\item{content}{The content of the message}

\item{role}{The role of the entity creating the message. Currently must be
'user'}

\item{files}{A list of assistant_file objects, maximum of 10, that the message should include}

\item{metadata}{A named character vector of up to 16 metadata values, with
names (keys) maximum 64 characters long and values maximum 512 characters
long}

\item{thread}{GPT_thread object representing the thread in which the message
is to be created}
}
\description{
Add a message to a GPT_thread
}
