% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{add_message}
\alias{add_message}
\title{Generic function to add a message}
\usage{
add_message(x, content, role = "user")
}
\arguments{
\item{x}{Object to add message to}

\item{content}{The content of the message}

\item{role}{The role of the message (defaults to "user")}
}
\description{
Generic function to add a message
}
