% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_embedding.R
\name{align_embeddings}
\alias{align_embeddings}
\title{Enforce additional alignment of cell clusters beyond the direct differential embedding}
\usage{
align_embeddings(fit, alignment = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{alignment}{a factor of length \code{ncol(data)}. The method tries to put elements with the
the same factor level close to each other. \code{NA} entries are ignored.}

\item{verbose}{Should the method print information during the fitting. Default: \code{TRUE}.}

\item{...}{additional parameters that are passed on to the internal function \code{align_points_impl}.
For example, \code{n_iter}, \code{tolerance}, and \code{target_layout}}
}
\description{
Enforce additional alignment of cell clusters beyond the direct differential embedding
}
