% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_impl}
\alias{align_impl}
\title{Align the points according to some grouping}
\usage{
align_impl(
  embedding,
  grouping,
  design_matrix,
  ridge_penalty = 0.01,
  preserve_position_of_NAs = FALSE,
  calculate_new_embedding = TRUE
)
}
\value{
A list with the new embedding and the coefficients
}
\description{
Align the points according to some grouping
}
\keyword{internal}
