% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_neighbors}
\alias{align_neighbors}
\alias{align_harmony}
\alias{align_by_template}
\alias{align_by_grouping}
\title{Enforce additional alignment of cell clusters beyond the direct differential embedding}
\usage{
align_neighbors(
  fit,
  method = c("rotation", "stretching", "rotation+stretching"),
  data_matrix = assay(fit),
  cells_per_cluster = 20,
  mnn = 10,
  design = fit$alignment_design_matrix,
  ridge_penalty = 0,
  verbose = TRUE
)

align_harmony(
  fit,
  method = c("rotation", "stretching", "rotation+stretching"),
  ...,
  design = fit$alignment_design_matrix,
  ridge_penalty = 0,
  min_cluster_membership = 0.1,
  max_iter = 10,
  verbose = TRUE
)

align_by_template(
  fit,
  method = c("rotation", "stretching", "rotation+stretching"),
  alignment_template,
  cells_per_cluster = 20,
  mnn = 10,
  design = fit$alignment_design_matrix,
  ridge_penalty = 0,
  verbose = TRUE
)

align_by_grouping(
  fit,
  method = c("rotation", "stretching", "rotation+stretching"),
  grouping,
  design = fit$alignment_design_matrix,
  ridge_penalty = 0,
  verbose = TRUE
)
}
\arguments{
\item{fit}{a \code{DiffEmbSeqFit} object}

\item{method}{how flexible should the alignment be}

\item{verbose}{Should the method print information during the fitting. Default: \code{TRUE}.}

\item{...}{additional parameters that are passed on to relevant functions}

\item{design_matrix}{the design matrix for the alignment}
}
\description{
Enforce additional alignment of cell clusters beyond the direct differential embedding
}
