% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_regression.R
\name{analytic_approx_procrustes_spd}
\alias{analytic_approx_procrustes_spd}
\title{Solve ||Y - P Z ||^2_2 for P in SPD(n) (aka. Symmetric Positive Definite Procrustes analysis)}
\usage{
analytic_approx_procrustes_spd(data, obs_embedding)
}
\description{
Solve ||Y - P Z ||^2_2 for P in SPD(n) (aka. Symmetric Positive Definite Procrustes analysis)
}
\keyword{internal}
