% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{The application server-side}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for shiny.}
}
\description{
The application server-side
}
\keyword{internal}
