% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{as.data.frame.chat}
\alias{as.data.frame.chat}
\title{Convert chat object to data frame}
\usage{
\method{as.data.frame}{chat}(x, ...)
}
\arguments{
\item{x}{A chat object}

\item{...}{Other arguments to be passed to data.frame()}
}
\description{
Convert chat object to data frame
}
