% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{as.vector.chat}
\alias{as.vector.chat}
\title{Convert chat object to a human-readable vector}
\usage{
\method{as.vector}{chat}(x, mode = "any")
}
\arguments{
\item{x}{A chat object}

\item{mode}{For compatibility with as.vector(); ignored}
}
\description{
Convert chat object to a human-readable vector
}
