% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant-class.R
\name{as_assistant.response}
\alias{as_assistant.response}
\title{An as_assistant method for httr responses}
\usage{
\method{as_assistant}{response}(response)
}
\arguments{
\item{response}{The httr response}
}
\description{
An as_assistant method for httr responses
}
