% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_file-class.R
\name{as_assistant_file.response}
\alias{as_assistant_file.response}
\title{An as_assistant_file method for httr responses}
\usage{
\method{as_assistant_file}{response}(response)
}
\arguments{
\item{response}{The httr response}
}
\description{
An as_assistant_file method for httr responses
}
