% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{as_chat.data.frame}
\alias{as_chat.data.frame}
\title{Convert data frame to chat object}
\usage{
\method{as_chat}{data.frame}(df = data.frame())
}
\arguments{
\item{df}{A data frame with columns 'role' and 'content' of character type}
}
\description{
Convert data frame to chat object
}
