% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{as_chat}
\alias{as_chat}
\title{Generic method to coerce an object to a chat}
\usage{
as_chat(x)
}
\arguments{
\item{x}{The object to coerce}
}
\description{
Generic method to coerce an object to a chat
}
