% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-class.R
\name{as_message.response}
\alias{as_message.response}
\title{An as_message method for httr responses}
\usage{
\method{as_message}{response}(response)
}
\arguments{
\item{response}{The httr response}
}
\description{
An as_message method for httr responses
}
