% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_file-class.R
\name{as_message_file.response}
\alias{as_message_file.response}
\title{An as_message_file method for httr responses}
\usage{
\method{as_message_file}{response}(response)
}
\arguments{
\item{response}{The httr response}
}
\description{
An as_message_file method for httr responses
}
