% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-class.R
\name{as_run.response}
\alias{as_run.response}
\title{An as_run method for httr responses}
\usage{
\method{as_run}{response}(response)
}
\arguments{
\item{response}{The httr response}
}
\description{
An as_run method for httr responses
}
