% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_step-class.R
\name{as_run_step.response}
\alias{as_run_step.response}
\title{An as_run_step method for httr responses}
\usage{
\method{as_run_step}{response}(response)
}
\arguments{
\item{response}{The httr response}
}
\description{
An as_run_step method for httr responses
}
