% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread-class.R
\name{as_thread}
\alias{as_thread}
\title{An as_thread method for httr responses}
\usage{
as_thread(response)
}
\arguments{
\item{response}{The httr response}
}
\description{
An as_thread method for httr responses
}
