% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_file-class.R
\name{assistant_file}
\alias{assistant_file}
\title{Helper to create assistant_file objects}
\usage{
assistant_file(id = NULL, created_at = NULL, assistant_id = NULL)
}
\arguments{
\item{type}{The type of file, one of "retrieval" (the default), "code_interpreter", or "function"}
}
\description{
Helper to create assistant_file objects
}
