% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-utils.R
\name{box}
\alias{box}
\title{shinydashboard::box - function}
\usage{
box(
  ...,
  title = NULL,
  footer = NULL,
  status = NULL,
  solidHeader = FALSE,
  background = NULL,
  width = 6,
  height = NULL,
  collapsible = FALSE,
  collapsed = FALSE
)
}
\description{
shinydashboard::box - function
}
\keyword{internal}
