% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-modify_LT.R
\name{build_reduction_matrix}
\alias{build_reduction_matrix}
\title{Build reduction matrix to be used in the app}
\usage{
build_reduction_matrix(data, select_cod, select_x, cod_change)
}
\description{
Build reduction matrix to be used in the app
}
\keyword{internal}
