% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-runs.R
\name{cancel_run}
\alias{cancel_run}
\title{Cancel a run}
\usage{
cancel_run(thread_id, run_id)
}
\arguments{
\item{thread_id}{The ID of the thread the run belongs to (required)}

\item{run_id}{The ID of the run to cancel (required)}
}
\value{
The modified run object
}
\description{
Cancel a run
}
\references{
\url{https://platform.openai.com/docs/api-reference/runs/cancelRun}
}
