% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffEmbFit.R
\name{.DollarNames.DiffEmbFit}
\alias{.DollarNames.DiffEmbFit}
\alias{$,DiffEmbFit-method}
\alias{dollar_methods}
\alias{$<-,DiffEmbFit-method}
\title{Fluent use of accessor methods}
\usage{
\method{.DollarNames}{DiffEmbFit}(x, pattern = "")

\S4method{$}{DiffEmbFit}(x, name)

\S4method{$}{DiffEmbFit}(x, name) <- value
}
\description{
Fluent use of accessor methods
}
\seealso{
\link{accessor_methods} for more documentation on the
accessor functions.
}
