% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur_fit.R
\name{.DollarNames.lemur_fit}
\alias{.DollarNames.lemur_fit}
\alias{$,lemur_fit-method}
\alias{dollar_methods}
\alias{$<-,lemur_fit-method}
\title{Access values from a \code{lemur_fit}}
\usage{
\method{.DollarNames}{lemur_fit}(x, pattern = "")

\S4method{$}{lemur_fit}(x, name)

\S4method{$}{lemur_fit}(x, name) <- value
}
\arguments{
\item{x}{the \code{lemur_fit}}

\item{pattern}{the pattern from looking up potential values interactively}

\item{name}{the name of the value behind the dollar}

\item{value}{the replacement value. This only works for \code{colData} and
\code{rowData}.}
}
\description{
Access values from a \code{lemur_fit}
}
\seealso{
\code{\linkS4class{lemur_fit}} for more documentation on the
accessor functions.
}
