% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur_fit_obj.R
\name{.DollarNames.lemur_fit_obj}
\alias{.DollarNames.lemur_fit_obj}
\alias{$,lemur_fit_obj-method}
\alias{dollar_methods}
\alias{$<-,lemur_fit_obj-method}
\title{Fluent use of accessor methods}
\usage{
\method{.DollarNames}{lemur_fit_obj}(x, pattern = "")

\S4method{$}{lemur_fit_obj}(x, name)

\S4method{$}{lemur_fit_obj}(x, name) <- value
}
\description{
Fluent use of accessor methods
}
\seealso{
\link{accessor_methods} for more documentation on the
accessor functions.
}
