% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_based_ensemble.R
\name{cdf_stacking_forecaster_v1pre}
\alias{cdf_stacking_forecaster_v1pre}
\title{Form a cdf-stacking v1pre forecast}
\usage{
cdf_stacking_forecaster_v1pre(
  response,
  incidence_period,
  ahead,
  geo_type,
  n_locations,
  component_forecaster_names,
  panterpolate_cdf,
  n_training_dates = 3L,
  n_required_instances_per_forecaster = 10L,
  prerange.constant.extension = 1,
  prerange.proportional.extension = 0.2,
  shared_naming_scheme = character(0L),
  repo_root_dirpath = get_repo_root_dirpath(),
  historical_naming_scheme = shared_naming_scheme,
  prospective_naming_scheme = shared_naming_scheme,
  hub_naming_scheme = shared_naming_scheme,
  historical_components_dirpath = get_historical_components_dirpath(repo_root_dirpath),
 
    prospective_components_dirpath = get_prospective_components_dirpath(repo_root_dirpath),
  hub_components_dirpath = get_hub_components_dirpath(repo_root_dirpath)
)
}
\arguments{
\item{response}{length-1 character vector, optionally named; the response for
which to generate a forecast; this ensemble is formed from components with
a response matching either the content or name of this argument}

\item{incidence_period}{length-1 character vector; the
\code{incidence_period} for which to generate a forecast; this ensemble is
formed from components with the same \code{incidence_period}}

\item{ahead}{length-1 integer-valued \code{is.numeric}; the \code{ahead} for
which to generate a forecast; this ensemble is formed from components with
the same \code{ahead}}

\item{geo_type}{length-1 character vector; the \code{geo_type} for which to
generate a forecast; this ensemble is formed from components with the same
\code{geo_type}}

\item{n_locations}{length-1 integer-valued \code{is.numeric}; the
\code{n_locations} for which to generate a forecast; this ensemble is
formed from components with the same \code{n_locations}}

\item{component_forecaster_names}{names of forecasters to try to include}

\item{shared_naming_scheme}{(re)naming scheme used as default for historical,
prospective, and hub forecasters; see details}

\item{repo_root_dirpath}{length-1 character vector; path to root of the
covidcast-forecast repository working directory from which to read
component forecasts; defaults to the lowest ancestor of the \code{getwd}
with a .git file/directory}

\item{historical_naming_scheme}{(re)naming scheme --- override for
shared_naming_scheme if needed for historical_components_dirpath}

\item{prospective_naming_scheme}{(re)naming scheme --- override for
shared_naming_scheme if needed for prospective_components_dirpath}

\item{hub_naming_scheme}{(re)naming scheme --- override for
shared_naming_scheme if needed for hub_components_dirpath}

\item{historical_components_dirpath}{override for the path to the directory
(tree) containing historical component prediction/score cards}

\item{prospective_components_dirpath}{override for the path to the directory
(tree) containing prospective component prediction/score cards}

\item{hub_components_dirpath}{override for the path to the directory (tree)
containing COVID-19 Forecasting Hub component prediction/score cards}

\item{forecast_date}{ignored (to completely discover whether an ensemble can
be formed requires reading card files to find whether they are NA, but this
makes \code{\link{quantileEnsemble::get_forecasters}} run slowly;
currently, no checks are done on the "outer" \code{forecast_date} fed to
this function, and only the \code{forecast_date} fed to a forecaster
function output from this function is considered)}
}
\description{
Form a cdf-stacking v1pre forecast
}
\details{
Each (re)naming scheme is a character vector: each named vector
entry denotes renaming forecasters with the entry's name to the entry
itself; each unnamed vector entry denotes a name that is expected to be
found somewhere in the searched paths; each entry (named or unnamed) is
checked for appearance somewhere in any of the searched paths. However, the
converse is not checked, i.e., it is fine for there to be paths found that
do not correspond to an entry in the naming schemes.
}
