% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama-service.R
\name{check_ollama_response}
\alias{check_ollama_response}
\title{Check the status code of a response from the ollama API}
\usage{
check_ollama_response(response)
}
\arguments{
\item{response}{The response from the httr call}
}
\description{
Check the status code of a response from the ollama API
}
