% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-helpers.R
\name{check_openai_response}
\alias{check_openai_response}
\title{Check the status code of a response from the OpenAI API}
\usage{
check_openai_response(response)
}
\arguments{
\item{response}{The response from the httr call}
}
\description{
Check the status code of a response from the OpenAI API
}
