% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{complete_GPT.chat}
\alias{complete_GPT.chat}
\title{Create chat completions with the OpenAI API}
\usage{
\method{complete_GPT}{chat}(x)
}
\arguments{
\item{x}{A chat object}
}
\description{
API endpoint is documented at
\url{https://platform.openai.com/docs/api-reference/chat/create}
}
