% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{complete_GPT}
\alias{complete_GPT}
\title{Generic method to complete an object such as a chat with GPT}
\usage{
complete_GPT(x)
}
\arguments{
\item{x}{The object}
}
\description{
Generic method to complete an object such as a chat with GPT
}
