% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{complete_GPT_tryCatch}
\alias{complete_GPT_tryCatch}
\title{tryCatch-buffered complete_GPT}
\usage{
complete_GPT_tryCatch(x, tries = 3)
}
\arguments{
\item{x}{A chat object}

\item{tries}{How many times to try before giving up}
}
\description{
tryCatch-buffered complete_GPT
}
