% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{correct_design_matrix_groups}
\alias{correct_design_matrix_groups}
\title{Find the rotation and stretching coefficients to make latent position of matching groups similar}
\usage{
correct_design_matrix_groups(
  fit,
  matching_groups,
  embedding,
  design,
  rotating = TRUE,
  stretching = TRUE,
  ridge_penalty = 0,
  verbose = TRUE
)
}
\arguments{
\item{matching_groups}{a list with two (plus one optional) elements.
\describe{
\item{matches}{a list of vectors with the indices of the cells that are part of the match}
\item{index_groups}{a list of vectors whose lengths correspond to \code{match}. The entries correspond to
the separate conditions in the design matrix. \emph{This is weird because it is redundant with the \code{design} argument.}}
\item{weights}{an optional list of vectors whose lengths correspond to \code{match}. The entries correspond to
the weight of that cell in the match.}
}}
}
\description{
Find the rotation and stretching coefficients to make latent position of matching groups similar
}
\keyword{internal}
