% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-assistants.R
\name{create_assistant}
\alias{create_assistant}
\title{Create an assistant}
\usage{
create_assistant(
  name = NULL,
  description = NULL,
  instructions = NULL,
  tools = NULL,
  file_ids = NULL,
  metadata = NULL
)
}
\arguments{
\item{name}{The name of the assistant, an atomic character vector of maximum
length 256 characters}

\item{description}{A description of the assistant, an atomic character
vector of maximum length 512 characters}

\item{instructions}{Instructions for the assistant, an atomic character
vector of maximum length 32,768 characters}

\item{tools}{A list of assistant_tool objects, maximum of 128}

\item{file_ids}{A vector of file_ids to be attached to the assistant, maximum of 20}

\item{metadata}{A named character vector of up to 16 metadata values, with
names (keys) maximum 64 characters long and values maximum 512 characters
long}
}
\value{
An assistant object.
}
\description{
The model will be set from the OPENAI_MODEL environmental variable. All
other parameters are optional.
}
\references{
\url{https://platform.openai.com/docs/api-reference/assistants/createAssistant}
}
