% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-assistants.R
\name{create_assistant_file}
\alias{create_assistant_file}
\title{Create an assistant file}
\usage{
create_assistant_file(assistant_id, file_id)
}
\arguments{
\item{assistant_id}{The ID of the assistant}

\item{file_id}{The ID of the file}
}
\value{
An assistant_file object.
}
\description{
Create an assistant file
}
\references{
\url{https://platform.openai.com/docs/api-reference/assistants/createAssistantFile}
}
