% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-runs.R
\name{create_run}
\alias{create_run}
\title{Create a run}
\usage{
create_run(
  thread_id,
  assistant_id,
  instructions = NULL,
  additional_instructions = NULL,
  tools = NULL,
  metadata = NULL,
  temperature = 1,
  stream = FALSE
)
}
\arguments{
\item{thread_id}{The ID of the thread to run (required)}

\item{assistant_id}{The ID of the assistant to be used to execute the run (required)}

\item{instructions}{Instructions for the assistant, an atomic character
vector of maximum length 32,768 characters. If provided, will override any
existing instructions for the assistant.}

\item{additional_instructions}{Additional instructions specific for the run.
An atomic character vector of maximum length 32,768 characters. Will not
override any existing instructions for the assistant.}

\item{tools}{A list of assistant_tool objects, maximum of 128}

\item{metadata}{A named character vector of up to 16 metadata values, with
names (keys) maximum 64 characters long and values maximum 512 characters
long}

\item{temperature}{The sampling temperature to be used for the run, a double
between 0 and 2. Defaults to 1.}

\item{stream}{Logical value; if TRUE, will return a stream. Defaults to
FALSE.}
}
\value{
A run object
}
\description{
The model will be set from the OPENAI_MODEL environmental variable.
}
\references{
\url{https://platform.openai.com/docs/api-reference/runs/createRun}#'
}
