% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-threads.R
\name{create_thread}
\alias{create_thread}
\title{Create a thread.}
\usage{
create_thread(messages = NULL, metadata = NULL)
}
\arguments{
\item{messages}{A list of message objects, to start the thread with (optional)}

\item{metadata}{A named character vector of up to 16 metadata values, with
names (keys) maximum 64 characters long and values maximum 512 characters
long}
}
\value{
A thread object
}
\description{
Create a thread.
}
\references{
\url{https://platform.openai.com/docs/api-reference/threads/createThread}
}
