% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaiassistant-service.R
\name{create_vector_store}
\alias{create_vector_store}
\title{Create a vector store with associated file(s), to use with the file_search
tool}
\usage{
create_vector_store(paths)
}
\arguments{
\item{paths}{List of paths to files to be attached to the vector store}
}
\description{
Create a vector store with associated file(s), to use with the file_search
tool
}
