% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityCauses-data.R
\docType{data}
\name{data_app_input}
\alias{data_app_input}
\title{Input data for the shiny app}
\format{
An object of class \code{list} of length 7.
}
\usage{
data_app_input
}
\description{
Input data for the shiny app
}
\keyword{datasets}
