% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityCauses-data.R
\docType{data}
\name{data_cod_mapping}
\alias{data_cod_mapping}
\title{Causes of Death List Mapped to ICD Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 61 rows and 7 columns.
}
\usage{
data_cod_mapping
}
\description{
This table contains the cause of death list used in the package 
mapped to International Classification of Diseases (ICD) codes: ICD-10, 
ICD-10 used in hospital/claim analyses, ICD-9 and ICD-9 used in 
hospital/claim analyses.
}
\examples{
data_cod_mapping
}
\keyword{datasets}
