% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityCauses-data.R
\docType{data}
\name{data_gbd2019}
\alias{data_gbd2019}
\title{Global Burden of Disease Study 2019 (GBD 2019) Results}
\format{
An object of class \code{list} of length 2.
}
\source{
Global Burden of Disease Collaborative Network.
Global Burden of Disease Study 2019 (GBD 2019) Results.
Seattle, United States: Institute for Health Metrics and Evaluation (IHME), 2020.
Available from http://ghdx.healthdata.org/gbd-results-tool.
}
\usage{
data_gbd2019
}
\description{
Global Burden of Disease Study 2019 (GBD 2019) Results
}
\keyword{datasets}
