% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur-data.R
\docType{data}
\name{data_gbd2019_cod}
\alias{data_gbd2019_cod}
\title{Causes of Death Data between 1990 and 2019 -- Global Burden of Disease Study 2019}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1739304 rows and 6 columns.
}
\source{
Global Burden of Disease Collaborative Network.
Global Burden of Disease Study 2019 (GBD 2019) Results.
Seattle, United States: Institute for Health Metrics and Evaluation (IHME), 2020.
Available from http://ghdx.healthdata.org/gbd-results-tool.
}
\usage{
data_gbd2019_cod
}
\description{
Causes of Death Data between 1990 and 2019 -- Global Burden of Disease Study 2019
}
\examples{
data_gbd2019_cod
}
\keyword{datasets}
