% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityCauses-data.R
\docType{data}
\name{data_gbd2019_lt}
\alias{data_gbd2019_lt}
\title{Abridged life table between 1990 and 2019 -- Global Burden of Disease Study 2019}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 102795 rows and 13 columns.
}
\source{
United Nations, Department of Economic and Social Affairs,
Population Division (2019). World Population Prospects 2019,
\href{http://ghdx.healthdata.org/record/ihme-data/gbd-2019-life-tables-1950-2019}{
 Online Edition. Rev. 1.}
}
\usage{
data_gbd2019_lt
}
\description{
Abridged life table between 1990 and 2019 -- Global Burden of Disease Study 2019
}
\examples{
data_gbd2019_lt
}
\keyword{datasets}
