% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityCauses-data.R
\docType{data}
\name{data_gbd2019_sdg}
\alias{data_gbd2019_sdg}
\title{Causes of Death Data for 2015-2020 -- Global Burden of Disease Study 2019}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1781453 rows and 7 columns.
}
\source{
Global Burden of Disease Collaborative Network.
Global Burden of Disease Study 2019 (GBD 2019) Results.
Seattle, United States: Institute for Health Metrics and Evaluation (IHME), 2020.
Available from http://ghdx.healthdata.org/gbd-results-tool.
}
\usage{
data_gbd2019_sdg
}
\description{
In this data object the death count data (GBD2019) is grouped in such a way 
that make possible the tracking of the evolution of the UN's
Sustainable Development Goals.
}
\examples{
data_gbd2019_sdg
}
\keyword{datasets}
