% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur-data.R
\name{data_gbd2021_lt}
\alias{data_gbd2021_lt}
\title{Abridged life table between 1990 and 2021 -- Global Burden of Disease Study 2022}
\source{
Global Burden of Disease Collaborative Network.
Seattle, United States: Institute for Health Metrics and Evaluation (IHME), 2022.
Available from https://vizhub.healthdata.org/gbd-results/.
\href{https://vizhub.healthdata.org/gbd-results/}{
 Global Burden of Disease Study 2021 (GBD 2021) Results}
}
\usage{
data_gbd2021_lt()
}
\value{
A \code{data.table} of abridged life tables: one row per region,
sex, period and age group. Columns: \code{region}, \code{period},
\code{sex}, \code{x.int}, \code{x}, \code{mx}, \code{qx}, \code{ax},
\code{lx}, \code{dx}, \code{Lx}, \code{Tx} and \code{ex}.
}
\description{
Life tables constructed using the GBD2021 probability of death data.
}
\details{
Returns the GBD2021 life tables used by the package, read from the package's
\file{inst/extdata/} directory as a pre-factorized \code{data.table}.
}
\examples{
L <- data_gbd2021_lt()
L[L$region == "Romania" & L$sex == "both" & L$period == 2021, ]
}
