% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityCauses-data.R
\docType{data}
\name{data_sf}
\alias{data_sf}
\title{Simple features data for the map}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 173 rows and 13 columns.
}
\usage{
data_sf
}
\description{
Simple features data for the map
}
\seealso{
\code{\link{plot_map}}
}
\keyword{datasets}
