% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityCauses-data.R
\docType{data}
\name{data_wpp2019_lt}
\alias{data_wpp2019_lt}
\title{Abridged life table for 2015-2020 -- World Population Prospects 2019}
\format{
An object of class \code{list} of length 2.
}
\source{
United Nations, Department of Economic and Social Affairs, 
Population Division (2019). World Population Prospects 2019,
\href{https://population.un.org/wpp/Download/Standard/Mortality/}{
 Online Edition. Rev. 1.}
}
\usage{
data_wpp2019_lt
}
\description{
Abridged life table for 2015-2020 -- World Population Prospects 2019
}
\examples{
data_wpp2019_lt
}
\keyword{datasets}
