% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-decompose.R
\name{decompose_by_cod}
\alias{decompose_by_cod}
\title{Perform decomposition of age- and cause-specific mortality contributions
in life expectancy between any two regions/time periods}
\source{
The code of this function is based on
\code{DemoDecomp::stepwise_replacement}, maintained by Tim Riffe.
}
\usage{
decompose_by_cod(L1, L2, C1, C2, symmetrical = TRUE, direction = "up")
}
\arguments{
\item{L1}{Life table corresponding to the region/time period 1.}

\item{L2}{Life table corresponding to the region/time period 2.}

\item{C1}{Causes of death data containing death counts corresponding
to the region/time period 1. Format: long table.}

\item{C2}{Causes of death data containing death counts corresponding
to the region/time period 2. Format: long table.}

\item{symmetrical}{logical. Whether to average the results of replacing 1
with 2 and 2 with 1. When FALSE, cod1 is replaced by cod2 only; when TRUE,
the arithmetic average of the replacement in both directions is used.
Defaults are set to symmetrically replace from the bottom up, per the
authors' suggestion.}

\item{direction}{character. One of "up", "down" or "both". Default "up",
as recommended by the authors. It refers to whether we replace vector
elements from the bottom up or top down, or take the arithmetic average of
these. Although the total difference will always sum correctly, the
contribution from individual components can vary with the order of
replacement.}
}
\value{
A long data.frame in the same format as the cod input data. The
column \code{decomposition} indicates the change in life expectancy by age
and/or cause of death between the two life tables provided as input.
Unit: years.
}
\description{
Perform decomposition of age- and cause-specific mortality contributions
in life expectancy between any two regions/time periods
}
\details{
This implements the algorithm described in Andreev et al. (2002),
with defaults set to approximate their recommendations for replacement
ordering and result averaging.
}
\examples{
L <- data_gbd2021_lt()  # life tables
D <- data_gbd2021_cod() # cod data

# Select two Life Tables
region1 = "Romania"
region2 = "Mexico"
sex_sel = "male"
year    = 2021

lt1 <- L[L$region == region1 & L$sex == sex_sel & L$period == year, ]
lt2 <- L[L$region == region2 & L$sex == sex_sel & L$period == year, ]

# Select COD corresponding data
cod1 <- D[D$region == region1 & D$sex == sex_sel & D$period == year, ]
cod2 <- D[D$region == region2 & D$sex == sex_sel & D$period == year, ]

## Example of decomposition by age and cause of death
dec  <- decompose_by_cod(L1 = lt1,
                         L2 = lt2,
                         C1 = cod1,
                         C2 = cod2)

dec

plot_decompose(dec)
}
\references{
Andreev EM, Shkolnikov VM, and Begun AZ. Algorithm for
decomposition of differences between aggregate demographic measures and its
application to life expectancies, healthy life expectancies,
parity-progression ratios and total fertility rates. Demographic Research 7:
499-522. 2002.
}
