% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogr-functions.R
\name{decompose_ex_by_age}
\alias{decompose_ex_by_age}
\title{Perform decomposition of age-specific mortality contributions 
in life expectancy between any two populations}
\usage{
decompose_ex_by_age(A, B)
}
\arguments{
\item{A}{Life table 1.}

\item{B}{Life table 2.}
}
\description{
Decompose the difference in life expectancy (at the smallest age) 
of two populations represented by their life tables in a given 
period of time.
}
\examples{
library(MortalityCauses)
library(dplyr)

# Select Life Table 1
L1 <- data_gbd2019_lt \%>\% 
  filter(region == "Romania",
         sex == "both",
         level == "median")

# Life Table 2
L2 <- data_gbd2019_lt \%>\% 
  filter(region == "Republic of Moldova",
         sex == "both",
         level == "median")

# Life expectancy at birth difference
delta_ex <- L2$ex[1] - L1$ex[1] # about 2 years diff

# Age decomposition
delta_ex_by_age <- decompose_ex_by_age(L1, L2)
}
\references{
...
}
\author{
...
}
