% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-assistants.R
\name{delete_assistant}
\alias{delete_assistant}
\title{Delete an assistant}
\usage{
delete_assistant(assistant_id)
}
\arguments{
\item{assistant_id}{The ID of the assistant}
}
\value{
Nothing
}
\description{
Delete an assistant
}
\references{
\url{https://platform.openai.com/docs/api-reference/assistants/deleteAssistant}
}
