% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-assistants.R
\name{delete_assistant_file}
\alias{delete_assistant_file}
\title{Delete an assistant file}
\usage{
delete_assistant_file(assistant_id, file_id)
}
\arguments{
\item{assistant_id}{The ID of the assistant}

\item{file_id}{The ID of the assistant file}
}
\value{
Nothing
}
\description{
Delete an assistant file
}
\references{
\url{https://platform.openai.com/docs/api-reference/assistants/deleteAssistantFile}
}
