% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-files.R
\name{delete_file}
\alias{delete_file}
\title{Delete a file}
\usage{
delete_file(file_id)
}
\arguments{
\item{file_id}{The ID of the file to delete}
}
\description{
Delete a file
}
\references{
\url{https://platform.openai.com/docs/api-reference/files/delete}
}
