% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-threads.R
\name{delete_thread}
\alias{delete_thread}
\title{Delete a thread}
\usage{
delete_thread(thread_id)
}
\arguments{
\item{thread_id}{The ID of the thread}
}
\value{
Nothing
}
\description{
Delete a thread
}
\references{
\url{https://platform.openai.com/docs/api-reference/threads/deleteThread}
}
