% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{dt_filter}
\alias{dt_filter}
\title{Filter dataset using data.table methods}
\usage{
dt_filter(data, mode, region1, region2, gender, year)
}
\description{
Filter dataset using data.table methods
}
\keyword{internal}
