% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_filters.R
\name{dt_filter_local}
\alias{dt_filter_local}
\title{Filter dataset using data.table methods
This is used when the app runs in local mode in R}
\usage{
dt_filter_local(data, mode, region1, region2, gender, year, db_pool)
}
\description{
Filter dataset using data.table methods
This is used when the app runs in local mode in R
}
\keyword{internal}
