% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_variance.R
\name{estimate_variance}
\alias{estimate_variance}
\title{Bootstrap the fit to get an estimate of the parameter variances}
\usage{
estimate_variance(
  fit,
  n_bootstrap_samples = 100,
  replicates_by = NULL,
  refit_ambient_pca = FALSE,
  refit_linear_model = TRUE,
  refit_differential_embedding = TRUE,
  refit_alignment = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fit}{the result produced by \code{\link{differential_embedding}}}

\item{n_bootstrap_samples}{how many bootstrap samples to produce. Note that
the function has a high startup cost, but each additional bootstrap is fast.}

\item{refit_ambient_pca}{Fitting the ambient PCA is often the slowest step when
fitting the differential embedding model. Thus, by default, it is set to \code{FALSE}
for the bootstrap samples.}

\item{refit_linear_model, refit_differential_embedding, refit_alignment}{additional
flags to set which part of the model are copied from the original fit and which are
bootstrapped. By default, they are all \code{TRUE}}

\item{verbose}{Should the method print information during the fitting. Default: \code{TRUE}.}
}
\description{
Bootstrap the fit to get an estimate of the parameter variances
}
