% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-decompose.R
\name{exFUN}
\alias{exFUN}
\title{Life expectancy function}
\usage{
exFUN(x, cod)
}
\arguments{
\item{x}{A vector of ages.}

\item{cod}{A matrix containing cause-specific death rates.}
}
\value{
Life expectancy at birth (or at the minimum age).
}
\description{
Life expectancy function
}
\keyword{internal}
