% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_captions.R
\name{generate_fig2_captions}
\alias{generate_fig2_captions}
\title{Generate Fig2 captions
All arguments are shiny inputs except the last 2 life tables}
\usage{
generate_fig2_captions(
  mode,
  region1,
  region2,
  fig2_x,
  perc,
  cod_change,
  cod_target,
  lt_initial,
  lt_final
)
}
\description{
Generate Fig2 captions
All arguments are shiny inputs except the last 2 life tables
}
\keyword{internal}
