% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_captions.R
\name{generate_fig3_captions}
\alias{generate_fig3_captions}
\title{Generate Fig3 captions
All arguments are shiny inputs}
\usage{
generate_fig3_captions(perc)
}
\description{
Generate Fig3 captions
All arguments are shiny inputs
}
\keyword{internal}
