% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_captions.R
\name{generate_fig4_captions}
\alias{generate_fig4_captions}
\title{Generate Fig4 captions
All arguments are shiny inputs}
\usage{
generate_fig4_captions(perc, fig4_dim)
}
\description{
Generate Fig4 captions
All arguments are shiny inputs
}
\keyword{internal}
