% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_captions.R
\name{generate_table_captions}
\alias{generate_table_captions}
\title{Generate captions for data tables
All arguments are shiny inputs}
\usage{
generate_table_captions(mode, region1, region2, time_slider, sex, cod_change)
}
\description{
Generate captions for data tables
All arguments are shiny inputs
}
\keyword{internal}
