% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-threads.R
\name{get_messages.GPT_thread}
\alias{get_messages.GPT_thread}
\title{Update a thread with the current list of messages}
\usage{
\method{get_messages}{GPT_thread}(x)
}
\arguments{
\item{x}{A GPT_thread object}
}
\description{
Update a thread with the current list of messages
}
