% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen-ensemble.R
\name{get_quantile_forecasts}
\alias{get_quantile_forecasts}
\title{Get array of quantile forecasts from \code{forecasters} on \code{train_dates}}
\usage{
get_quantile_forecasts(
  train_dates,
  forecasters,
  ahead = 1,
  geo_type = "county",
  incidence_period = "epiweek",
  n_locations = 200,
  response = "usafacts_deaths_incidence_num",
  repo_root_dirpath = get_repo_root_dirpath(),
  from_scorecard = TRUE,
  locations_considered = NULL
)
}
\arguments{
\item{locations_considered}{\code{NULL} or character vector; if \code{NULL},
does nothing; if a character vector, restricts output to locations in this
vector}
}
\value{
3-D  array: (date;location) x forecaster x quantile;  can be fed into
Ryan's quantgen package
}
\description{
Requires that prediction_cards or score_cards are in directory structure like
Naras's. If  \code{from_scorecard} is  \code{TRUE}, then load  forecasts from
scorecard.RDS  and  return  actual   as  well.  If  \code{from_scorecard}  is
\code{FALSE}, then load  forecasts from out.RDS (pred card)  and don't return
actual.
}
