% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant.R
\name{GPT_assistant}
\alias{GPT_assistant}
\title{Helper to create GPT_assistant objects}
\usage{
GPT_assistant(
  name = NULL,
  description = NULL,
  instructions = NULL,
  tools = NULL,
  files = NULL,
  metadata = NULL,
  .dry_run = FALSE
)
}
\arguments{
\item{name}{The name of the assistant, an atomic character vector of maximum
length 256 characters}

\item{description}{A description of the assistant, an atomic character
vector of maximum length 512 characters}

\item{instructions}{Instructions for the assistance, an atomic character
vector of maximum length 32,768 characters}

\item{tools}{A list of assistant_tool objects, maximum of 128}

\item{files}{A list of assistant_file objects, maximum of 20}

\item{metadata}{A named character vector of up to 16 metadata values, with
names (keys) maximum 64 characters long and values maximum 512 characters
long}

\item{.dry_run}{If TRUE, will merely pretend to call the OpenAI API}
}
\description{
The model will be set from the OPENAI_MODEL environmental variable. All
other parameters are optional.
}
\details{
As part of creating the (local) GPT_assistant object, a linked assistant
object will be created with the OpenAI API, unless .dry_run is set.

API documented at
\url{https://platform.openai.com/docs/api-reference/assistants/createAssistant}
}
