% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-threads.R
\name{GPT_thread}
\alias{GPT_thread}
\title{Helper to create GPT_thread objects}
\usage{
GPT_thread(messages = list(), metadata = NULL)
}
\arguments{
\item{messages}{A list of GPT_message objects (not GPT_messages objects) to
start the thread with}

\item{metadata}{A named character vector of up to 16 metadata values, with
names (keys) maximum 64 characters long and values maximum 512 characters
long}
}
\description{
As part of creating the (local) GPT_thread object, a linked thread entity
will be created with the OpenAI API.
}
\details{
API documented at
\url{https://platform.openai.com/docs/api-reference/threads/createThread}
}
