% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_regression.R
\name{grassmann_lm}
\alias{grassmann_lm}
\title{Solve ||Y - exp_p(V * x) Y ||^2_2 for V}
\usage{
grassmann_lm(data, design, base_point, tangent_regression = FALSE)
}
\value{
A three-dimensional array with the coefficients \code{V}.
}
\description{
Solve ||Y - exp_p(V * x) Y ||^2_2 for V
}
\keyword{internal}
