% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hayman-ensemble.R
\name{hayman_ensemble_v1_predcard_dt}
\alias{hayman_ensemble_v1_predcard_dt}
\title{Prepare a hayman v1 predcard using predcards for the same forecast_date}
\usage{
hayman_ensemble_v1_predcard_dt(
  predcards_dt,
  test_forecast_date,
  required_forecaster_names = character(0L),
  include_all_other_forecasters = TRUE,
  optional_forecaster_names = character(0L),
  allow_other_generation_dates = FALSE,
  allow_other_prediction_dates = FALSE
)
}
\arguments{
\item{predcards_dt}{a \code{predcards_dt}  containing only the predcards that
should be  incorporated in  the ensemble}

\item{test_forecast_date}{a length-1 \code{\link{Date}}-compatible object}
}
\description{
Differs from v0 in that it performs the required forecaster checks on a
per-location-forecast_time-task and complete-distribution basis vs. just a
per-forecast_time-task and card-presence basis.
}
