% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-class.R
\name{initialise.chat}
\alias{initialise.chat}
\title{Initialise a chat}
\usage{
\method{initialise}{chat}(chat)
}
\arguments{
\item{chat}{The chat object}
}
\description{
Performs whatever steps and checks are needed to initialise a chat for a
given service. Dispatches to service-specific subclass methods.
}
