% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaichat-service.R
\name{initialise.openaichat}
\alias{initialise.openaichat}
\title{Initialise an openaichat chat}
\usage{
\method{initialise}{openaichat}(chat)
}
\arguments{
\item{chat}{The chat object}
}
\description{
Initialise an openaichat chat
}
