% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-class.R, R/ollama-service.R,
%   R/openaichat-service.R
\name{initialise}
\alias{initialise}
\alias{initialise.chat}
\alias{initialise.ollama}
\alias{initialise.openaichat}
\title{Initialise a chat}
\usage{
initialise(chat, ...)

\method{initialise}{chat}(chat, .verbose = TRUE, ...)

\method{initialise}{ollama}(chat, .verbose = TRUE, ...)

\method{initialise}{openaichat}(chat, .verbose = TRUE, ...)
}
\arguments{
\item{chat}{The chat object}

\item{...}{Other service-specific arguments}

\item{.verbose}{If FALSE, progress messages will be suppressed}
}
\description{
Performs whatever steps and checks are needed to initialise a chat
for a given service. Dispatches to service-specific subclass
methods.
}
