% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{last_message.chat}
\alias{last_message.chat}
\title{Return the last message from a chat object}
\usage{
\method{last_message}{chat}(x)
}
\arguments{
\item{x}{A chat object}
}
\description{
Return the last message from a chat object
}
