% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{last_message}
\alias{last_message}
\title{Generic method to return the last message from an object like a chat}
\usage{
last_message(x)
}
\arguments{
\item{x}{The object}
}
\description{
Generic method to return the last message from an object like a chat
}
