% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama-service.R
\name{last_response.ollama}
\alias{last_response.ollama}
\title{Get the last response sent by the model in an ollama chat}
\usage{
\method{last_response}{ollama}(chat, ...)
}
\arguments{
\item{chat}{The chat}

\item{...}{Further arguments passed from other methods}
}
\description{
Get the last response sent by the model in an ollama chat
}
