% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaiassistant-service.R
\name{last_response.openaiassistant}
\alias{last_response.openaiassistant}
\title{Get the last response sent by the model in an openaiassistant chat}
\usage{
\method{last_response}{openaiassistant}(chat)
}
\arguments{
\item{chat}{The chat}
}
\description{
Get the last response sent by the model in an openaiassistant chat
}
