% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaichat-service.R
\name{last_response.openaichat}
\alias{last_response.openaichat}
\title{Get the last response sent by the model in an openaichat}
\usage{
\method{last_response}{openaichat}(chat)
}
\arguments{
\item{chat}{The chat}
}
\description{
Get the last response sent by the model in an openaichat
}
