% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-class.R
\name{last_response}
\alias{last_response}
\alias{last_response.chat}
\title{Get the last response sent by the model in a a chat}
\usage{
last_response(chat, ...)

\method{last_response}{chat}(chat, ...)
}
\arguments{
\item{chat}{The chat object}

\item{...}{Other service-specific arguments}
}
\description{
Get the last response sent by the model in a a chat
}
