% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur-package.R
\docType{package}
\name{lemur-package}
\alias{lemur}
\alias{lemur-package}
\title{lemur: R bindings for LLM APIs}
\description{
Provides R bindings for LLM APIs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wilkox/lemur}
  \item Report bugs at \url{https://github.com/wilkox/lemur/issues}
}

}
\author{
\strong{Maintainer}: David Wilkins \email{david@wilkox.org}

}
\keyword{internal}
