% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-assistants.R
\name{list_assistant_files}
\alias{list_assistant_files}
\title{List files associated with an assistant}
\usage{
list_assistant_files(
  assistant_id,
  limit = 20,
  order = "desc",
  before = NULL,
  after = NULL
)
}
\arguments{
\item{assistant_id}{The ID of the assistant}

\item{limit}{A limit on the number of files to be returned. Can be between
1-100, defaults to 20.}

\item{order}{Sort order for the created_at timestamp. Either 'asc' for
ascending or 'desc' (default) for descending.}

\item{before}{Return assistants before the named assistant id}

\item{after}{Return assistants after the named assistant id}
}
\value{
Data frame of files
}
\description{
List files associated with an assistant
}
\references{
\url{https://platform.openai.com/docs/api-reference/assistants/listAssistantFiles}
}
