% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API-files.R
\name{list_files}
\alias{list_files}
\title{List files belonging to the user's organisation}
\usage{
list_files(.dry_run = FALSE)
}
\arguments{
\item{.dry_run}{If TRUE, will not make an actual call to the GPT API}
}
\description{
List files belonging to the user's organisation
}
