% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant.R
\name{list_GPT_assistants}
\alias{list_GPT_assistants}
\title{List GPT assistants}
\usage{
list_GPT_assistants(limit = 20, order = "desc", before = NULL, after = NULL)
}
\arguments{
\item{order}{Sort order for the created_at timestamp. Either 'asc' for
ascending or 'desc' (default) for descending.}

\item{before}{Return assistants before the named assistant id}

\item{after}{Return assistants after the named assistant id}

\item{name}{limit A limit on the number of assistants to be returned. Can be
between 1-100, defaults to 20.}
}
\value{
List of GPT assistants as a data frame
}
\description{
List GPT assistants
}
\references{
\url{https://platform.openai.com/docs/api-reference/assistants/listAssistants}
}
