% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-messages.R
\name{list_message_files}
\alias{list_message_files}
\title{List message files}
\usage{
list_message_files(
  thread_id,
  message_id,
  limit = 20,
  order = "desc",
  after = NULL,
  before = NULL
)
}
\arguments{
\item{thread_id}{The ID of the thread (required)}

\item{message_id}{The ID of the message (required)}

\item{limit}{A limit on the number of assistants to be returned. Can be
between 1-100, defaults to 20.}

\item{order}{Sort order for the created_at timestamp. Either 'asc' for
ascending or 'desc' (default) for descending.}

\item{after}{Return assistants after the named assistant id}

\item{before}{Return assistants before the named assistant id}
}
\value{
Data frame of message files
}
\description{
List files associated with a message
}
\references{
\url{https://platform.openai.com/docs/api-reference/messages/listMessageFiles}
}
