% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-messages.R
\name{list_messages}
\alias{list_messages}
\title{List messages}
\usage{
list_messages(
  thread_id,
  limit = 20,
  order = "desc",
  after = NULL,
  before = NULL
)
}
\arguments{
\item{thread_id}{The ID of the thread to list messages for (required)}

\item{limit}{A limit on the number of assistants to be returned. Can be
between 1-100, defaults to 20.}

\item{order}{Sort order for the created_at timestamp. Either 'asc' for
ascending or 'desc' (default) for descending.}

\item{after}{Return assistants after the named assistant id}

\item{before}{Return assistants before the named assistant id}
}
\value{
Data frame of messages
}
\description{
List messages for a given thread
}
\references{
\url{https://platform.openai.com/docs/api-reference/messages/listMessages}
}
