% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-runs.R
\name{list_runs}
\alias{list_runs}
\title{List runs}
\usage{
list_runs(thread_id, limit = 20, order = "desc", before = NULL, after = NULL)
}
\arguments{
\item{thread_id}{The ID of the thread to list runs for (required)}

\item{limit}{A limit on the number of assistants to be returned. Can be
between 1-100, defaults to 20.}

\item{order}{Sort order for the created_at timestamp. Either 'asc' for
ascending or 'desc' (default) for descending.}

\item{before}{Return assistants before the named assistant id}

\item{after}{Return assistants after the named assistant id}
}
\value{
Data frame of runs
}
\description{
List runs
}
\references{
\url{https://platform.openai.com/docs/api-reference/runs/listRuns}
}
