% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_methods.R
\name{MAST_DE}
\alias{MAST_DE}
\title{An implementation of MAST DE method.}
\usage{
MAST_DE(sce, cellgroups, repgroups, freq_expressed = 0.05)
}
\arguments{
\item{sce}{a SingleCellExperiment object with raw count matrix. The CPM counts are computed in the funciton.}

\item{cellgroups}{a vector of group labels for cells. The cells consist of two different groups.}

\item{repgroups}{a vector of donor labels for cells.}

\item{freq_expressed}{a threshold for gene detection rate.}
}
\value{
A dataframe of MAST DE results with each row for a gene.
}
\description{
An implementation of MAST DE method.
}
\examples{
data(Bcells_sce)
MAST_DE(Bcells_sce, Bcells_sce$stim, Bcells_sce$ind)
}
