% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_file-class.R
\name{message_file}
\alias{message_file}
\title{Helper to create message_file objects}
\usage{
message_file(id = NULL, created_at = NULL, message_id = NULL)
}
\description{
Helper to create message_file objects
}
