% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama-service.R
\name{messages.ollama}
\alias{messages.ollama}
\title{Get messages from an ollama chat}
\usage{
\method{messages}{ollama}(chat, ...)
}
\arguments{
\item{chat}{The chat}

\item{...}{Further arguments passed from other methods}
}
\description{
Get messages from an ollama chat
}
