% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaiassistant-service.R
\name{messages.openaiassistant}
\alias{messages.openaiassistant}
\title{Get messages from an openaiassistant chat}
\usage{
\method{messages}{openaiassistant}(chat, ...)
}
\arguments{
\item{chat}{The chat}

\item{...}{Further arguments passed from other methods}
}
\description{
Get messages from an openaiassistant chat
}
