% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaichat-service.R
\name{messages.openaichat}
\alias{messages.openaichat}
\title{Get messages from an openaichat chat}
\usage{
\method{messages}{openaichat}(chat, ...)
}
\arguments{
\item{chat}{The chat}

\item{...}{Further arguments passed from other methods}
}
\description{
Get messages from an openaichat chat
}
