% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server_map.R
\name{mod_map_server}
\alias{mod_map_server}
\title{Server module: MAP}
\usage{
mod_map_server(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
Server module: MAP
}
