% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-assistants.R
\name{modify_assistant}
\alias{modify_assistant}
\title{Modify an assistant}
\usage{
modify_assistant(
  assistant_id,
  name = NULL,
  description = NULL,
  instructions = NULL,
  tools = NULL,
  metadata = NULL
)
}
\arguments{
\item{assistant_id}{The ID of the assistant to modify (required)}

\item{name}{The name of the assistant, an atomic character vector of maximum
length 256 characters}

\item{description}{A description of the assistant, an atomic character
vector of maximum length 512 characters}

\item{instructions}{Instructions for the assistance, an atomic character
vector of maximum length 32,768 characters}

\item{tools}{A list of assistant_tool objects, maximum of 128}

\item{metadata}{A named character vector of up to 16 metadata values, with
names (keys) maximum 64 characters long and values maximum 512 characters
long}
}
\value{
An assistant object.
}
\description{
The model will be set from the OPENAI_MODEL environmental variable.
}
\references{
\url{https://platform.openai.com/docs/api-reference/assistants/modifyAssistant}
}
