% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-modify_LT.R
\name{modify_cod_table}
\alias{modify_cod_table}
\title{Modify COD table by changing the cause of death associated risks}
\usage{
modify_cod_table(cod, cod_change)
}
\arguments{
\item{cod}{Causes of death matrix containing death counts corresponding
to the population and time period of the life table.}

\item{cod_change}{Numerical scalar, vector or matrix.
The changes to be applied to the causes of death rate, m[c,x],
in order to reduce or increase the mortality estimates given by the life
table. Accepted input: any value greater than -100. See examples.}
}
\value{
A long table with the same format as the input data
}
\description{
Modify COD table by changing the cause of death associated risks
}
\examples{
D <- data_gbd2021_cod() # cod data

# Select COD data
cod <- D[D$region == "Romania" & D$sex == "both" & D$period == 2021, ]
cod_change = -50

# Example 1:
# Modify all COD values by 50\%. This is trivial and not really needed.
modify_cod_table(cod, cod_change = -50)


# Example 2:
# Change the first cod by 1\%, second one with 2\% and so on until 17\%
modify_cod_table(cod, cod_change = 1:17)

# Example 3:
# Apply a specific change by cause and age
# Say, we want to decrease the cod's risk only between age 45 and 75
# with values between 24\% and 40\%.

# we have to build a matrix with 25 rows and 18 columns (AGEs x CODs)
# to indicate the change for each combination
M <- matrix(24:40, nrow = 25, ncol = 18, byrow = TRUE)
dimnames(M) <- list(unique(cod$x), unique(cod$cause_name))
M[!(rownames(M) \%in\% 45:75), ] <- 0
M

modify_cod_table(cod, cod_change = -M)
}
