% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-runs.R
\name{modify_run}
\alias{modify_run}
\title{Modify a run}
\usage{
modify_run(thread_id, run_id, metadata = NULL)
}
\arguments{
\item{thread_id}{The ID of the thread the run belongs to (required)}

\item{run_id}{The ID of the run to modify (required)}

\item{metadata}{A named character vector of up to 16 metadata values, with
names (keys) maximum 64 characters long and values maximum 512 characters
long}
}
\value{
The modified run object
}
\description{
Modify a run
}
\references{
\url{https://platform.openai.com/docs/api-reference/runs/modifyRun}
}
