% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-threads.R
\name{modify_thread}
\alias{modify_thread}
\title{Modify a thread}
\usage{
modify_thread(thread_id, metadata = NULL)
}
\arguments{
\item{thread_id}{The ID of the thread}

\item{metadata}{A named character vector of up to 16 metadata values, with
names (keys) maximum 64 characters long and values maximum 512 characters
long}
}
\value{
A thread object
}
\description{
Modify a thread
}
\references{
\url{https://platform.openai.com/docs/api-reference/threads/modifyThread}
}
