% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur-package.R
\docType{package}
\name{MortalityCauses}
\alias{lemur}
\alias{lemur-package}
\alias{MortalityCauses}
\title{lemur: Life expectancy monitor upscaled in R}
\description{
The life expectancy monitoring tool allows the user to selected mortality changes over the entire lifespan or at specific ages, as well as for overall mortality or for specific causes of death. For example, how would life expectancy look if cardiovascular mortality were to be reduced by 50%? Or how would life expectancy look if infant mortality was eliminated? The tool facilitates assessing changes and comparisons in life expectancy under those selected scenarios of mortality change. Furthermore, the tool lets the user compare cause-of-death profiles and life expectancies across time, countries and sexes.
}
\details{
To learn more about the package, start with the vignettes:
\code{browseVignettes(package = "lemur")}
}
\author{
\strong{Maintainer}: Marius D. Pascariu \email{rpascariu@outlook.com} (\href{https://orcid.org/0000-0002-2568-6489}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Marius D. Pascariu \email{rpascariu@outlook.com} (\href{https://orcid.org/0000-0002-2568-6489}{ORCID}) [copyright holder]
  \item Jose Manuel Aburto (\href{https://orcid.org/0000-0002-2926-6879}{ORCID})
  \item Vladimir Canudas-Romo (\href{https://orcid.org/0000-0001-6532-0089}{ORCID})
}

}
