% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mply_dbl}
\alias{mply_dbl}
\alias{msply_dbl}
\alias{stack_rows}
\alias{stack_cols}
\title{Iterating function that returns a matrix}
\usage{
mply_dbl(x, FUN, ncol = 1, ...)

msply_dbl(x, FUN, ...)

stack_rows(x)

stack_cols(x)
}
\description{
The length of \code{x} determines the number of rows. The length of
\code{FUN(x[i])} determines the number of columns. Must match \code{ncol}.
}
\section{Functions}{
\itemize{
\item \code{msply_dbl}: flexible version that automatically infers the number
of columns

\item \code{stack_rows}: Each list element becomes a row in a matrix

\item \code{stack_cols}: Each list element becomes a row in a matrix
}}

\keyword{internal}
