% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant-class.R
\name{new_assistant}
\alias{new_assistant}
\title{A constructor function for the assistant class}
\usage{
new_assistant(x = list())
}
\arguments{
\item{x}{A list containing parameters for the assistant, as documented in
assistant()}
}
\description{
A constructor function for the assistant class
}
