% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_file-class.R
\name{new_assistant_file}
\alias{new_assistant_file}
\title{A constructor function for the assistant_file class}
\usage{
new_assistant_file(x = list())
}
\arguments{
\item{x}{A list containing parameters for the assistant file, as documented in
assistant_file()}
}
\description{
A constructor function for the assistant_file class
}
