% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_tool-class.R
\name{new_assistant_tool}
\alias{new_assistant_tool}
\title{A constructor function for the assistant_tool class}
\usage{
new_assistant_tool(x = list())
}
\arguments{
\item{x}{A list containing parameters for the assistant tool, as documented in
assistant_tool()}
}
\description{
A constructor function for the assistant_tool class
}
