% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-class.R
\name{new_chat}
\alias{new_chat}
\title{A constructor function for chat objects}
\usage{
new_chat(x = list(service = character(), model = character()))
}
\arguments{
\item{x}{A list containing parameters for the chat, including at least
\code{service} and \code{model}}
}
\description{
A constructor function for chat objects
}
