% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant.R
\name{new_GPT_assistant}
\alias{new_GPT_assistant}
\title{A constructor function for GPT_assistant}
\usage{
new_GPT_assistant(x = list())
}
\arguments{
\item{x}{A list containing parameters for the assistant, as documented in
GPT_assistant()}
}
\description{
A constructor function for GPT_assistant
}
