% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-threads.R
\name{new_GPT_thread}
\alias{new_GPT_thread}
\title{A constructor function for GPT_thread}
\usage{
new_GPT_thread(x = list())
}
\arguments{
\item{x}{A list containing parameters for the thread, as documented in
GPT_thread()}
}
\description{
A constructor function for GPT_thread
}
