% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-class.R
\name{new_message}
\alias{new_message}
\title{A constructor function for message objects}
\usage{
new_message(x = list())
}
\arguments{
\item{x}{A list containing parameters for the message}
}
\description{
A constructor function for message objects
}
