% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_file-class.R
\name{new_message_file}
\alias{new_message_file}
\title{A constructor function for the message_fileclass}
\usage{
new_message_file(x = list())
}
\arguments{
\item{x}{A list containing parameters for the message file, as documented in
message_file()}
}
\description{
A constructor function for the message_fileclass
}
