% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-class.R
\name{new_run}
\alias{new_run}
\title{A constructor function for the run class}
\usage{
new_run(x = list())
}
\arguments{
\item{x}{A list containing parameters for the run, as documented in
run()}
}
\description{
A constructor function for the run class
}
