% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_step-class.R
\name{new_run_step}
\alias{new_run_step}
\title{A constructor function for the run_step class}
\usage{
new_run_step(x = list())
}
\arguments{
\item{x}{A list containing parameters for the run step, as documented in
run_step()}
}
\description{
A constructor function for the run_step class
}
