% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread-class.R
\name{new_thread}
\alias{new_thread}
\title{A constructor function for thread objects}
\usage{
new_thread(x = list())
}
\arguments{
\item{x}{A list containing parameters for the thread}
}
\description{
A constructor function for thread objects
}
