% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-utils.R
\name{number}
\alias{number}
\title{scales::number (vendored)}
\usage{
number(
  x,
  accuracy = NULL,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = " ",
  decimal.mark = ".",
  trim = TRUE,
  ...
)
}
\description{
scales::number (vendored)
}
\keyword{internal}
