% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_based_ensemble.R
\name{offset_cdf_inner_product}
\alias{offset_cdf_inner_product}
\title{Calculates \int_{-\infty}^{+\infty} \link{F_1(z)F_2(z) - I(z > 0)} dz}
\usage{
offset_cdf_inner_product(cdf1, cdf2, fallback, ...)
}
\description{
Calculates \int_{-\infty}^{+\infty} \link{F_1(z)F_2(z) - I(z > 0)} dz
}
