% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{one_hot_encoding}
\alias{one_hot_encoding}
\title{Take a vector and convert it to a one-hot encoded matrix}
\usage{
one_hot_encoding(groups)
}
\value{
A matrix with \code{length(unique(groups))} rows and \code{length(groups)} columns.
}
\description{
Take a vector and convert it to a one-hot encoded matrix
}
\keyword{internal}
