% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-api.R
\name{openai_api_key}
\alias{openai_api_key}
\title{Retrieve OpenAI API key from the OPENAI_API_KEY environmental variable}
\usage{
openai_api_key()
}
\description{
Retrieve OpenAI API key from the OPENAI_API_KEY environmental variable
}
