% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_based_ensemble.R
\name{panterpolate_reweight2_jump_cdf}
\alias{panterpolate_reweight2_jump_cdf}
\title{Get jump CDF from quantiles}
\usage{
panterpolate_reweight2_jump_cdf(quantiles, probs)
}
\arguments{
\item{quantiles}{vector of quantiles}

\item{probs}{the levels of these quantiles}
}
\description{
Fits a piecewise constant CDF from the quantiles using approxfun, using
\code{\link{quantile_level_sample_weights}} to handle ambiguity in a "symmetric" way
at the low and high tails and make this correspond to a somewhat natural pdf.
}
