% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-figures.R
\name{plot_cod}
\alias{plot_cod}
\title{Plot function for COD data}
\usage{
plot_cod(cod, perc = FALSE, type = "barplot")
}
\arguments{
\item{cod}{Causes of death matrix containing death counts corresponding
to the population and time period of the life table.}

\item{perc}{Logical. If TRUE data will be displayed as percentages else
as absolute values. Default: FALSE.}

\item{type}{Options: "barplot" or "piechart".}
}
\value{
A plotly widget.
}
\description{
Plot function for COD data
}
\examples{
D <- data_gbd2021_cod() # cod data
cod <- D[D$region == "Romania" & D$sex == "both" & D$period == 2021, ]
plot_cod(cod)
}
