% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-figures.R
\name{plot_decompose}
\alias{plot_decompose}
\title{Plot a decomposition object}
\usage{
plot_decompose(object, perc = FALSE, by = "both")
}
\arguments{
\item{object}{An object of class \code{decompose}.}

\item{perc}{Logical. If TRUE data will be displayed as percentages else
as absolute values. Default: FALSE.}

\item{by}{The dimensions on which to build the plot.
One of "both", "age" or "cod".}
}
\value{
A plotly widget.
}
\description{
Plot a decomposition object
}
\examples{
# See example in the ?decompose_by_cod or ?decompose_by_age help pages
}
\seealso{
\code{\link{decompose_by_cod}}
\code{\link{decompose_by_age}}
}
