% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-figures.R
\name{plot_map}
\alias{plot_map}
\title{Plot an interactive map}
\usage{
plot_map(location, zoom = 5, data = lemur::data_sf)
}
\arguments{
\item{location}{Geographical location.}

\item{zoom}{The zoom level.}

\item{data}{An \code{sf} object with the map polygons.
Default: \code{lemur::data_sf}.}
}
\value{
A leaflet widget.
}
\description{
Plot an interactive map
}
\examples{
plot_map(location = "Mexico")

}
