% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-figures.R
\name{plotly_change}
\alias{plotly_change}
\title{Native plotly version of plot_change (app internal)}
\usage{
plotly_change(
  L1,
  L2,
  age = seq(0, 110, by = 10),
  perc = FALSE,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{L1}{Life table corresponding to the region/time period 1.}

\item{L2}{Life table corresponding to the region/time period 2.}

\item{age}{Reference ages.}

\item{perc}{Logical. If TRUE data will be displayed as percentages else
as absolute values. Default: FALSE.}

\item{xlab, ylab}{Axis titles. If \code{NULL} the same defaults as
\code{\link{plot_change}} are used.}
}
\description{
Native plotly version of plot_change (app internal)
}
\keyword{internal}
