% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-figures.R
\name{plotly_cod}
\alias{plotly_cod}
\title{Native plotly version of plot_cod (app internal)}
\usage{
plotly_cod(cod, perc = FALSE, xlab = NULL, mode = "mode_cod")
}
\arguments{
\item{cod}{Causes of death matrix containing death counts corresponding
to the population and time period of the life table.}

\item{perc}{Logical. If TRUE data will be displayed as percentages else
as absolute values. Default: FALSE.}

\item{xlab}{X axis title. If \code{NULL} the same defaults as
\code{\link{plot_cod}} are used.}

\item{mode}{One of "mode_cod", "mode_cntr", "mode_sex", "mode_sdg",
"mode_sdg2" - passed through from the app to decide whether the data
spans two regions or two sexes and needs a two-panel subplot (the
native equivalent of facet_wrap()).}
}
\description{
Native plotly version of plot_cod (app internal)
}
\keyword{internal}
