% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-figures.R
\name{plotly_decompose}
\alias{plotly_decompose}
\title{Native plotly version of plot_decompose (app internal)}
\usage{
plotly_decompose(
  object,
  perc = FALSE,
  by = "both",
  xlab = NULL,
  ylab = NULL,
  ttip = c("fill", "y")
)
}
\arguments{
\item{object}{An object of class \code{decompose}.}

\item{perc}{Logical. If TRUE data will be displayed as percentages else
as absolute values. Default: FALSE.}

\item{by}{The dimensions on which to build the plot.
One of "both", "age" or "cod".}

\item{xlab, ylab}{Axis titles. If \code{NULL} the same defaults as
\code{\link{plot_decompose}} are used.}

\item{ttip}{Hover/tooltip fields, as returned by the app caption system
(e.g. \code{c("fill", "y")}). Used to select hoverinfo.}
}
\description{
Native plotly version of plot_decompose (app internal)
}
\keyword{internal}
