% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_methods.R
\name{poisson_glmm_DE}
\alias{poisson_glmm_DE}
\title{An implementation Poisson glmm DE method.}
\usage{
poisson_glmm_DE(sce, cellgroups, repgroups, freq_expressed = 0.05)
}
\arguments{
\item{sce}{a SingleCellExperiment object with raw count matrix.}

\item{cellgroups}{a vector of group labels for cells. The cells consist of two different groups.}

\item{repgroups}{a vector of donor labels for cells.}

\item{freq_expressed}{a threshold for gene detection rate.}
}
\value{
A dataframe of Poisson glmm DE results with each row for a gene.
}
\description{
An implementation Poisson glmm DE method.
}
\examples{
data(Bcells_sce)
poisson_glmm_DE(Bcells_sce, Bcells_sce$stim, Bcells_sce$ind)
}
