% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.DiffEmbFit}
\alias{predict.DiffEmbFit}
\title{Predict values from \code{DiffEmbFit} object}
\usage{
\method{predict}{DiffEmbFit}(
  object,
  newdata = NULL,
  newdesign = NULL,
  diffemb_embedding = object$diffemb_embedding,
  with_ambient_pca = TRUE,
  with_linear_model = TRUE,
  with_differential_embedding = TRUE,
  with_alignment = TRUE,
  ...
)
}
\description{
Predict values from \code{DiffEmbFit} object
}
