% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lemur_fit_obj}
\alias{predict.lemur_fit_obj}
\title{Predict values from \code{lemur_fit_obj} object}
\usage{
\method{predict}{lemur_fit_obj}(
  object,
  newdata = NULL,
  newdesign = NULL,
  embedding = object$embedding,
  with_ambient_pca = TRUE,
  with_linear_model = TRUE,
  with_differential_embedding = TRUE,
  with_alignment = TRUE,
  ...
)
}
\description{
Predict values from \code{lemur_fit_obj} object
}
