% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{prepare_data_mode1}
\alias{prepare_data_mode1}
\title{Prepare data for risk changes}
\usage{
prepare_data_mode1(cod, lt, region1, region2, cod_target, cod_change)
}
\description{
Prepare data for risk changes
}
\keyword{internal}
