% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{prepare_data_mode2}
\alias{prepare_data_mode2}
\title{Prepare data for country comparisons}
\usage{
prepare_data_mode2(cod, lt, region1, region2, cod_target, cod_change)
}
\description{
Prepare data for country comparisons
}
\keyword{internal}
