% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_process_data.R
\name{prepare_data_mode_sdg}
\alias{prepare_data_mode_sdg}
\title{Prepare data for risk changes}
\usage{
prepare_data_mode_sdg(
  cod,
  lt,
  region1,
  cod_target,
  sdg_1,
  sdg_3,
  sdg_4,
  sdg_5,
  sdg_6,
  sdg_7
)
}
\description{
Prepare data for risk changes
}
\keyword{internal}
