% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_processing.R
\name{prepare_data_mode_sex}
\alias{prepare_data_mode_sex}
\title{Prepare data for sex comparisons}
\usage{
prepare_data_mode_sex(cod, lt, region1, region2, sex, cod_change)
}
\description{
Prepare data for sex comparisons
}
\keyword{internal}
