% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_file-class.R
\name{print.assistant_file}
\alias{print.assistant_file}
\title{A print method for the assistant_file class}
\usage{
\method{print}{assistant_file}(x, ...)
}
\arguments{
\item{x}{An assistant_file object}

\item{...}{Other arguments to be passed to print()}
}
\description{
A print method for the assistant_file class
}
