% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_tool-class.R
\name{print.assistant_tool}
\alias{print.assistant_tool}
\title{A print method for the assistant_tool class}
\usage{
\method{print}{assistant_tool}(x, ...)
}
\arguments{
\item{x}{An assistant_tool object}

\item{...}{Other arguments to be passed to print()}
}
\description{
A print method for the assistant_tool class
}
