% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{print.chat}
\alias{print.chat}
\title{A print method for chat}
\usage{
\method{print}{chat}(x, ...)
}
\arguments{
\item{x}{A chat object}

\item{...}{Other arguments to be passed to print()}
}
\description{
Coloured output based on
\url{https://github.com/r-lib/testthat/blob/717b02164def5c1f027d3a20b889dae35428b6d7/R/colour-text.r}
}
