% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant.R
\name{print.GPT_assistant}
\alias{print.GPT_assistant}
\title{A print method for GPT_assistant}
\usage{
\method{print}{GPT_assistant}(x, ...)
}
\arguments{
\item{x}{A GPT_assistant object}

\item{...}{Other arguments to be passed to print()}
}
\description{
Coloured output based on
\url{https://github.com/r-lib/testthat/blob/717b02164def5c1f027d3a20b889dae35428b6d7/R/colour-text.r}
}
