% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-threads.R
\name{print.GPT_thread}
\alias{print.GPT_thread}
\title{A print method for GPT_thread}
\usage{
\method{print}{GPT_thread}(x, get_messages = TRUE, ...)
}
\arguments{
\item{x}{A GPT_thread object}

\item{get_messages}{Whether to get an up-to-date list of messages in the
thread prior from the OpenAI API prior to printing. TRUE by default}

\item{...}{Other arguments to be passed to print()}
}
\description{
Coloured output based on
\url{https://github.com/r-lib/testthat/blob/717b02164def5c1f027d3a20b889dae35428b6d7/R/colour-text.r}
}
