% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-class.R
\name{print.message}
\alias{print.message}
\title{A print method for message objects}
\usage{
\method{print}{message}(x, ...)
}
\arguments{
\item{x}{A message object}

\item{...}{Other arguments to be passed to print()}
}
\description{
A print method for message objects
}
