% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_file-class.R
\name{print.message_file}
\alias{print.message_file}
\title{A print method for the message_file class}
\usage{
\method{print}{message_file}(x, ...)
}
\arguments{
\item{x}{A message_file object}

\item{...}{Other arguments to be passed to print()}
}
\description{
A print method for the message_file class
}
