% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama-service.R
\name{print.ollama}
\alias{print.ollama}
\title{Print information about an ollama chat}
\usage{
\method{print}{ollama}(x, ...)
}
\arguments{
\item{x}{The chat}

\item{...}{Further arguments passed from other methods}
}
\description{
Print information about an ollama chat
}
