% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaiassistant-service.R
\name{print.openaiassistant}
\alias{print.openaiassistant}
\title{Print information about an openaiassistant chat}
\usage{
\method{print}{openaiassistant}(x, ...)
}
\arguments{
\item{x}{The chat}

\item{...}{Further arguments passed from other methods}
}
\description{
Print information about an openaiassistant chat
}
