% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaichat-service.R
\name{print.openaichat}
\alias{print.openaichat}
\title{Print information about an openaichat chat}
\usage{
\method{print}{openaichat}(chat)
}
\arguments{
\item{chat}{The chat}
}
\description{
Print information about an openaichat chat
}
