% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-class.R
\name{print.run}
\alias{print.run}
\title{A print method for the run class}
\usage{
\method{print}{run}(x, ...)
}
\arguments{
\item{x}{A run object}

\item{...}{Other arguments to be passed to print()}
}
\description{
A print method for the run class
}
