% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread-class.R
\name{print.thread}
\alias{print.thread}
\title{A print method for thread objects}
\usage{
\method{print}{thread}(x, ...)
}
\arguments{
\item{x}{A thread object}

\item{...}{Other arguments to be passed to print()}
}
\description{
A print method for thread objects
}
