% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_regression.R
\name{procrustes_rotation}
\alias{procrustes_rotation}
\title{Solve ||Y - R Z ||^2_2 for R in SO(n) (aka. Procrustes analysis)}
\usage{
procrustes_rotation(Y, Z)
}
\description{
Solve ||Y - R Z ||^2_2 for R in SO(n) (aka. Procrustes analysis)
}
