% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_regression.R
\name{procrustes_spd}
\alias{procrustes_spd}
\title{Solve ||B - A X ||^2_2 for A in SPD(n) (aka. Symmetric Positive Definite Procrustes analysis)}
\usage{
procrustes_spd(data, obs_embedding, maxiter = 1000, tolerance = 1e-08)
}
\value{
The matrix \code{A}.
}
\description{
Solve ||B - A X ||^2_2 for A in SPD(n) (aka. Symmetric Positive Definite Procrustes analysis)
}
\keyword{internal}
