% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileEnsemble-package.R
\docType{package}
\name{quantileEnsemble-package}
\alias{quantileEnsemble-package}
\title{quantileEnsemble is a forecaster package}
\author{
first_name last_name
Maintainer: first_name last_name<email>
}
\keyword{package}
