% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hayman-ensemble.R
\name{read_cards_dt}
\alias{read_cards_dt}
\title{Read prediction or score cards for all times and components}
\usage{
read_cards_dt(
  candidate_paths_dt,
  response,
  incidence_period,
  ahead,
  forecast_dates,
  geo_type,
  n_locations,
  expect_found_forecaster_names = character(0L),
  only_read_forecaster_names = NULL
)
}
\description{
All character arguments are meant to be actual values rather than bash globs.
\code{forecast_dates} can hold multiple values,  while others are expected to
be single values.
}
\examples{
read_forecasts_dt("jhu-csse_deaths_incidence_num","epiweek",2L,"2020-04-06","state",200L,
                  get_historical_components_dirpath(), c("strawman"))

read_forecasts_dt("jhu-csse_deaths_incidence_num","epiweek",2L,"2020-06-22","state",200L,
                  get_historical_components_dirpath(), c("aardvark"="aardvark_vanilla","aardvark_vanilla","poiszero"))

read_forecasts_dt("jhu-csse_deaths_incidence_num","epiweek",2L,"2020-04-06","state",200L,
                  get_historical_components_dirpath(), c("bogus1"="bogus1_renamed","bogus2","bogus2_altname"="bogus2"))

read_forecasts_dt("jhu-csse_deaths_incidence_num","epiweek",2L,"2020-06-22","state",200L,
                  get_prospective_components_dirpath(), c("aardvark"="aardvark_vanilla","aardvark_vanilla","poiszero"))

read_forecasts_dt("jhu-csse_deaths_incidence_num","epiweek",2L,"2020-06-22","state",200L,
                  get_hub_components_dirpath(), character(0L),
                  file_basename="scorecard.RDS")

}
