% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogr-functions.R
\name{reduce_LifeTable}
\alias{reduce_LifeTable}
\title{Modify life table by changing the cause of death associated risks}
\usage{
reduce_LifeTable(lt, cod, cod_change)
}
\arguments{
\item{lt}{Life table;}

\item{cod}{Causes of death matrix corresponding to the population and 
time period of the life table;}

\item{cod_change}{The change to be applied to the causes of death risks
in order to reduce of increase the mortality estimates given by the life 
table.}
}
\description{
Modify life table by changing the cause of death associated risks
}
\examples{
L <- data_gbd2019_lt  # life tables
D <- data_gbd2019_cod # cod data

# Select Life Table
lt <- L[L$region == "Romania" & L$sex == "both" & L$level == "median", ]
# Select COD data
cod <- D[D$region == "Romania" & D$sex == "both" & D$level == "median", ]

# How does the life table modify if the cause-specific mortality is
# reduced by 50\%?
lt_reduced <- reduce_LifeTable(lt, cod, cod_change = 50)
lt_reduced
}
