% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{remove_message.chat}
\alias{remove_message.chat}
\title{Remove the last message from a chat object}
\usage{
\method{remove_message}{chat}(x)
}
\arguments{
\item{x}{A chat object}
}
\description{
Remove the last message from a chat object
}
