% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{residuals,DiffEmbFit-method}
\alias{residuals,DiffEmbFit-method}
\title{Predict values from \code{DiffEmbFit} object}
\usage{
\S4method{residuals}{DiffEmbFit}(
  object,
  with_linear_model = TRUE,
  with_differential_embedding = TRUE,
  ...
)
}
\description{
Predict values from \code{DiffEmbFit} object
}
