% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{residuals,lemur_fit_obj-method}
\alias{residuals,lemur_fit_obj-method}
\title{Predict values from \code{lemur_fit_obj} object}
\usage{
\S4method{residuals}{lemur_fit_obj}(
  object,
  with_linear_model = TRUE,
  with_differential_embedding = TRUE,
  ...
)
}
\description{
Predict values from \code{lemur_fit_obj} object
}
