% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-assistants.R
\name{retrieve_assistant}
\alias{retrieve_assistant}
\title{Retrieve an assistant}
\usage{
retrieve_assistant(assistant_id)
}
\arguments{
\item{assistant_id}{The ID of the assistant}
}
\value{
An assistant object
}
\description{
Retrieve an assistant
}
\references{
\url{https://platform.openai.com/docs/api-reference/assistants/getAssistant}
}
