% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-assistants.R
\name{retrieve_assistant_file}
\alias{retrieve_assistant_file}
\title{Retrieve an assistant file}
\usage{
retrieve_assistant_file(assistant_id, file_id)
}
\arguments{
\item{assistant_id}{The ID of the assistant}

\item{file_id}{The ID of the assistant file}
}
\value{
An assistant_file object
}
\description{
Retrieve an assistant file
}
\references{
\url{https://platform.openai.com/docs/api-reference/assistants/getAssistantFile}
}
