% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-messages.R
\name{retrieve_message}
\alias{retrieve_message}
\title{Retrieve a message}
\usage{
retrieve_message(thread_id, message_id)
}
\arguments{
\item{thread_id}{The ID of the thread (required)}

\item{message_id}{The ID of the message (required)}
}
\value{
A message object
}
\description{
Retrieve a message
}
\references{
\url{https://platform.openai.com/docs/api-reference/messages/getMessage}
}
