% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-messages.R
\name{retrieve_message_file}
\alias{retrieve_message_file}
\title{Retrieve a message file}
\usage{
retrieve_message_file(thread_id, message_id, file_id)
}
\arguments{
\item{thread_id}{The ID of the thread (required)}

\item{message_id}{The ID of the message (required)}

\item{file_id}{The ID of the file (required)}
}
\value{
A message_file object
}
\description{
Retrieve a message file
}
\references{
\url{https://platform.openai.com/docs/api-reference/messages/getMessageFile}
}
