% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-runs.R
\name{retrieve_run}
\alias{retrieve_run}
\title{Retrieve a run}
\usage{
retrieve_run(thread_id, run_id)
}
\arguments{
\item{thread_id}{The ID of the thread the run belongs to (required)}

\item{run_id}{The ID of the run to retrieve (required)}
}
\value{
A run object
}
\description{
Retrieve a run
}
\references{
\url{https://platform.openai.com/docs/api-reference/runs/getRun}
}
