% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-runs.R
\name{retrieve_run_step}
\alias{retrieve_run_step}
\title{Retrieve a run step}
\usage{
retrieve_run_step(thread_id, run_id, step_id)
}
\arguments{
\item{thread_id}{The ID of the thread the run belongs to (required)}

\item{run_id}{The ID of the run the step belongs to (required)}

\item{step_id}{The ID of the step to retrieve (required)}
}
\value{
A run step object
}
\description{
Retrieve a run step
}
\references{
\url{https://platform.openai.com/docs/api-reference/runs/getRunStep}
}
