% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-threads.R
\name{retrieve_thread}
\alias{retrieve_thread}
\title{Retrieve a thread}
\usage{
retrieve_thread(thread_id)
}
\arguments{
\item{thread_id}{The ID of the thread}
}
\value{
A thread object
}
\description{
Retrieve a thread
}
\references{
\url{https://platform.openai.com/docs/api-reference/threads/getThread}
}
