% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge regression}
\usage{
ridge_regression(Y, X, ridge_penalty = 0, weights = rep(1, nrow(X)))
}
\arguments{
\item{Y}{the observations matrix (\verb{features x samples})}

\item{X}{the design matrix (\verb{samples x covariates})}

\item{ridge_penalty}{a numeric vector or matrix of size (\code{covariates} or
\verb{covariates x covariates} respectively)}

\item{weights}{a vector of observation weights}
}
\description{
The function does not treat the intercept special.
}
\keyword{internal}
