% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_regression.R
\name{rotation_geodesic_regression}
\alias{rotation_geodesic_regression}
\title{Solve d(R, exp_p(V * x))^2 for V}
\usage{
rotation_geodesic_regression(
  rotations,
  design,
  base_point,
  weights = 1,
  ridge_penalty = 0,
  tangent_regression = FALSE
)
}
\description{
Solve d(R, exp_p(V * x))^2 for V
}
