% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaiassistant-service.R
\name{say.openaiassistant}
\alias{say.openaiassistant}
\title{Send a message in an openaiassistant chat}
\usage{
\method{say}{openaiassistant}(chat, content)
}
\arguments{
\item{chat}{The chat}

\item{content}{The message content. Text only}
}
\description{
Send a message in an openaiassistant chat
}
