% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaichat-service.R
\name{say.openaichat}
\alias{say.openaichat}
\title{Send a message in an openaichat chat}
\usage{
\method{say}{openaichat}(chat, content, role = "user")
}
\arguments{
\item{chat}{The chat}

\item{content}{The message content. Text only}

\item{role}{The role, one of "user" (default), "system", or "assistant"}
}
\description{
Send a message in an openaichat chat
}
