% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-class.R, R/openaiassistant-service.R,
%   R/openaichat-service.R
\name{say}
\alias{say}
\alias{say.chat}
\alias{say.openaiassistant}
\alias{say.openaichat}
\title{Send a message in a chat}
\usage{
say(chat, content, ...)

\method{say}{chat}(chat, content, ...)

\method{say}{openaiassistant}(chat, content, respond = TRUE)

\method{say}{openaichat}(chat, content, role = "user")
}
\arguments{
\item{chat}{The chat}

\item{content}{The message content. Text only}

\item{...}{Other service-specific arguments}

\item{respond}{If TRUE (the default), after sending the message the
assistant will be run to generate a response. Setting \code{respond = FALSE}
allows sending multiple messages to the assistant before it is asked to
respond}

\item{role}{The role, one of "user" (default) or "system"}
}
\description{
Send a message in a chat
}
