% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{say_GPT.chat}
\alias{say_GPT.chat}
\title{A convenience function to send a message to GPT and capture the response}
\usage{
\method{say_GPT}{chat}(x, content, role = "user")
}
\arguments{
\item{x}{A chat object}

\item{content}{The content of the message}

\item{role}{The role of the message (defaults to "user")}
}
\description{
A convenience function to send a message to GPT and capture the response
}
