% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_methods.R
\name{simple_mean_DE}
\alias{simple_mean_DE}
\title{simple_mean_DE(counts(Bcells_sce), Bcells_sce$stim == "ctrl", Bcells_sce$stim == "stim")}
\usage{
simple_mean_DE(counts, cellgroup1, cellgroup2)
}
\description{
simple_mean_DE(counts(Bcells_sce), Bcells_sce$stim == "ctrl", Bcells_sce$stim == "stim")
}
