% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_regression.R
\name{spd_geodesic_regression}
\alias{spd_geodesic_regression}
\title{Solve d(P, exp_p(V * x))^2 for V}
\usage{
spd_geodesic_regression(
  spd_matrices,
  design,
  base_point,
  weights = 1,
  ridge_penalty = 0,
  tangent_regression = FALSE
)
}
\description{
Solve d(P, exp_p(V * x))^2 for V
}
\keyword{internal}
