% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stack_slice}
\alias{stack_slice}
\alias{destack_slice}
\title{Make a cube from a list of matrices}
\usage{
stack_slice(x)

destack_slice(x)
}
\description{
Make a cube from a list of matrices
}
\section{Functions}{
\itemize{
\item \code{destack_slice}: Make a list of matrices from a cube
}}

