% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaichat-service.R
\name{transcript.openaichat}
\alias{transcript.openaichat}
\title{The transcript of an openaichat chat}
\usage{
\method{transcript}{openaichat}(chat)
}
\arguments{
\item{chat}{The chat}
}
\description{
The transcript of an openaichat chat
}
