% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-class.R
\name{transcript}
\alias{transcript}
\alias{transcript.chat}
\title{Transcript of messages in a chat}
\usage{
transcript(chat, transcript, ...)

\method{transcript}{chat}(chat, ...)
}
\arguments{
\item{chat}{The chat object}

\item{...}{Other service-specific arguments}
}
\value{
A data frame
}
\description{
Transcript of messages in a chat
}
