% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hayman-ensemble.R
\name{untrained_qspace_ensemble_v1_predcard_dt}
\alias{untrained_qspace_ensemble_v1_predcard_dt}
\title{Prepare an untrained quantile-space ensemble v1 predcard_dt}
\usage{
untrained_qspace_ensemble_v1_predcard_dt(
  predcards_dt,
  test_forecast_date,
  ensemble_function,
  required_forecaster_names = character(0L),
  include_all_other_forecasters = TRUE,
  optional_forecaster_names = character(0L),
  allow_other_generation_dates = FALSE,
  allow_other_prediction_dates = FALSE
)
}
\arguments{
\item{predcards_dt}{a \code{predcards_dt}  containing only the predcards that
should be  incorporated in  the ensemble}

\item{test_forecast_date}{a length-1 \code{\link{Date}}-compatible object}

\item{ensemble_function}{a function, such as \code{\link{median}}, that
combines quantile predictions from multiple forecasters for a single
level&location&date&ahead into a single ensemble quantile prediction}
}
\description{
Prepare an untrained quantile-space ensemble v1 predcard_dt
}
