% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API-files.R
\name{upload_file}
\alias{upload_file}
\title{Upload a file}
\usage{
upload_file(file, purpose = "assistants", .dry_run = FALSE)
}
\arguments{
\item{file}{The file to upload}

\item{purpose}{The purpose of the file, usually either 'assistants' or 'fine-tuning'}

\item{.dry_run}{If TRUE, will not make an actual call to the GPT API}
}
\description{
Upload a file
}
