% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-assistants.R
\name{validate_assistant}
\alias{validate_assistant}
\title{A validator function for assistant objects}
\usage{
validate_assistant(x)
}
\arguments{
\item{x}{A assistant object}
}
\description{
A validator function for assistant objects
}
