% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_file-class.R
\name{validate_assistant_file}
\alias{validate_assistant_file}
\title{A validator function for the assistant_file class}
\usage{
validate_assistant_file(x)
}
\arguments{
\item{x}{A assistant_file object}
}
\description{
A validator function for the assistant_file class
}
