% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_tool-class.R
\name{validate_assistant_tool}
\alias{validate_assistant_tool}
\title{A validator function for the assistant_tool class}
\usage{
validate_assistant_tool(x)
}
\arguments{
\item{x}{A assistant_tool object}
}
\description{
A validator function for the assistant_tool class
}
