% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant.R
\name{validate_GPT_assistant}
\alias{validate_GPT_assistant}
\title{A validator function for GPT_assistant}
\usage{
validate_GPT_assistant(x)
}
\arguments{
\item{x}{A GPT_assistant object}
}
\description{
A validator function for GPT_assistant
}
