% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-threads.R
\name{validate_GPT_thread}
\alias{validate_GPT_thread}
\title{A validator function for GPT_thread}
\usage{
validate_GPT_thread(x)
}
\arguments{
\item{x}{A GPT_thread object}
}
\description{
A validator function for GPT_thread
}
