% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-class.R
\name{validate_message}
\alias{validate_message}
\title{A validator function for message objects}
\usage{
validate_message(x)
}
\arguments{
\item{x}{A message object}
}
\description{
A validator function for message objects
}
