% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_file-class.R
\name{validate_message_file}
\alias{validate_message_file}
\title{A validator function for the message_file class}
\usage{
validate_message_file(x)
}
\arguments{
\item{x}{A message_file object}
}
\description{
A validator function for the message_file class
}
