% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_step-class.R
\name{validate_run_step}
\alias{validate_run_step}
\title{A validator function for the run_step class}
\usage{
validate_run_step(x)
}
\arguments{
\item{x}{A run_step object}
}
\description{
A validator function for the run_step class
}
