% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread-class.R
\name{validate_thread}
\alias{validate_thread}
\title{A validator function for thread objects}
\usage{
validate_thread(x)
}
\arguments{
\item{x}{A thread object}
}
\description{
A validator function for thread objects
}
