% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-utils.R
\name{visible_dt}
\alias{visible_dt}
\title{Get "visible" version of a data.table}
\usage{
visible_dt(DT)
}
\value{
the data.table with its internal invisibility flag unset
}
\description{
\code{data.table} objects  have an internal version  of invisibility separate
from  \code{base::invisible}, which  disables \code{print}  output the  first
time that they are printed after  certain operations (e.g., mutating to add a
column).  This function ensures that internal invisibility flag is unset.
}
