% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hayman-ensemble.R
\name{weighted_quantile_type1}
\alias{weighted_quantile_type1}
\title{Calculates a weighted type-1 quantiles}
\usage{
weighted_quantile_type1(x, w, probs, na.rm = FALSE)
}
\arguments{
\item{x}{instance values}

\item{w}{instance importance weights}

\item{probs}{quantile levels to calculate}
}
\value{
quantile values corresponding to levels \code{probs}
}
\description{
Calculates a weighted type-1 quantiles
}
\seealso{
quantile
}
