% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur_fit_obj.R
\name{accessor_methods}
\alias{accessor_methods}
\alias{n_ambient,lemur_fit_obj-method}
\alias{n_embedding,lemur_fit_obj-method}
\alias{ambient_coordsystem,lemur_fit_obj-method}
\alias{ambient_offset,lemur_fit_obj-method}
\alias{design,lemur_fit_obj-method}
\alias{diffemb_basepoint,lemur_fit_obj-method}
\alias{diffemb_coefficients,lemur_fit_obj-method}
\alias{diffemb_embedding,lemur_fit_obj-method}
\alias{alignment_method,lemur_fit_obj-method}
\alias{alignment_rotation,lemur_fit_obj-method}
\alias{alignment_stretching,lemur_fit_obj-method}
\alias{alignment_design,lemur_fit_obj-method}
\alias{alignment_design_matrix,lemur_fit_obj-method}
\alias{design_matrix,lemur_fit_obj-method}
\alias{linear_coefficients,lemur_fit_obj-method}
\alias{bootstrap_samples,lemur_fit_obj-method}
\alias{knn_graph,lemur_fit_obj-method}
\title{Get different features and elements of the 'lemur_fit_obj' object}
\usage{
\S4method{n_ambient}{lemur_fit_obj}(object)

\S4method{n_embedding}{lemur_fit_obj}(object)

\S4method{ambient_coordsystem}{lemur_fit_obj}(object)

\S4method{ambient_offset}{lemur_fit_obj}(object)

\S4method{design}{lemur_fit_obj}(object)

\S4method{diffemb_basepoint}{lemur_fit_obj}(object)

\S4method{diffemb_coefficients}{lemur_fit_obj}(object)

\S4method{diffemb_embedding}{lemur_fit_obj}(object)

\S4method{alignment_method}{lemur_fit_obj}(object)

\S4method{alignment_rotation}{lemur_fit_obj}(object)

\S4method{alignment_stretching}{lemur_fit_obj}(object)

\S4method{alignment_design}{lemur_fit_obj}(object)

\S4method{alignment_design_matrix}{lemur_fit_obj}(object)

\S4method{design_matrix}{lemur_fit_obj}(object)

\S4method{linear_coefficients}{lemur_fit_obj}(object)

\S4method{bootstrap_samples}{lemur_fit_obj}(object)

\S4method{knn_graph}{lemur_fit_obj}(object)
}
\arguments{
\item{object}{the 'lemur_fit_obj' object}
}
\value{
See the documentation of the generics to find out what each method returns
}
\description{
The functions listed below can all be accessed using the
fluent dollar notation (ie. \code{fit$n_ambient}) without
any additional parentheses.
}
