% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffEmbFit.R
\name{accessor_methods}
\alias{accessor_methods}
\alias{n_ambient,DiffEmbFit-method}
\alias{n_embedding,DiffEmbFit-method}
\alias{ambient_coordsystem,DiffEmbFit-method}
\alias{ambient_offset,DiffEmbFit-method}
\alias{design,DiffEmbFit-method}
\alias{diffemb_basepoint,DiffEmbFit-method}
\alias{diffemb_coefficients,DiffEmbFit-method}
\alias{diffemb_embedding,DiffEmbFit-method}
\alias{alignment_method,DiffEmbFit-method}
\alias{alignment_coefficients,DiffEmbFit-method}
\alias{design_matrix,DiffEmbFit-method}
\alias{linear_coefficients,DiffEmbFit-method}
\alias{bootstrap_samples,DiffEmbFit-method}
\title{Get different features and elements of the 'DiffEmbFit' object}
\usage{
\S4method{n_ambient}{DiffEmbFit}(object)

\S4method{n_embedding}{DiffEmbFit}(object)

\S4method{ambient_coordsystem}{DiffEmbFit}(object)

\S4method{ambient_offset}{DiffEmbFit}(object)

\S4method{design}{DiffEmbFit}(object)

\S4method{diffemb_basepoint}{DiffEmbFit}(object)

\S4method{diffemb_coefficients}{DiffEmbFit}(object)

\S4method{diffemb_embedding}{DiffEmbFit}(object)

\S4method{alignment_method}{DiffEmbFit}(object)

\S4method{alignment_coefficients}{DiffEmbFit}(object)

\S4method{design_matrix}{DiffEmbFit}(object)

\S4method{linear_coefficients}{DiffEmbFit}(object)

\S4method{bootstrap_samples}{DiffEmbFit}(object)
}
\arguments{
\item{object}{the 'DiffEmbFit' object}
}
\value{
See the documentation of the generics to find out what each method returns
}
\description{
The functions listed below can all be accessed using the
fluent dollar notation (ie. \code{fit$n_ambient}) without
any additional parentheses.
}
