% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_tool-class.R
\name{assistant_tool}
\alias{assistant_tool}
\title{Helper to create assistant_tool objects}
\usage{
assistant_tool(type = "file_search")
}
\arguments{
\item{type}{The type of tool, one of "file_search" (the default), "code_interpreter", or "function"}
}
\description{
Helper to create assistant_tool objects
}
