% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-modify_LT.R
\name{build_cod_matrix}
\alias{build_cod_matrix}
\title{Format COD data}
\usage{
build_cod_matrix(cod)
}
\arguments{
\item{cod}{COD long table}
}
\value{
A matrix with percentages.
}
\description{
Transform the COD count data from a long table to a matrix
containing percentages, with ages as rows and CODs as columns.
}
\examples{
# cod data
D <- data_gbd2021_cod()
# Select COD data for 1 region
cod <- D[D$region == "Romania" & D$sex == "both" & D$period == 1990, ]
# COD data in matrix format

build_cod_matrix(cod)
}
