% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogr-functions.R
\name{build_cod_matrix}
\alias{build_cod_matrix}
\title{Transform the COD data from a long table to a matrix with ages as rows and 
cod's as columns.}
\usage{
build_cod_matrix(cod)
}
\arguments{
\item{cod}{COD long table}
}
\value{
A matrix with percentages.
}
\description{
Transform the COD data from a long table to a matrix with ages as rows and 
cod's as columns.
}
\examples{
# cod data
D <- data_gbd2019_cod 
# Select COD data for 1 region
cod <- D[D$region == "Romania" & D$sex == "both" & D$level == "median", ]
# COD data in matrix format

build_cod_matrix(cod) 
}
