% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogr-functions.R
\name{build_cod_matrix}
\alias{build_cod_matrix}
\title{Transform the COD data from a long table to a matrix with ages as rows and 
cod's as columns.}
\usage{
build_cod_matrix(cod, na.rm = TRUE)
}
\arguments{
\item{cod}{COD long table}

\item{na.rm}{logical.Should missing values be replaced with zero's?}
}
\value{
A matrix with percentages.
}
\description{
Transform the COD data from a long table to a matrix with ages as rows and 
cod's as columns.
}
\keyword{internal}
