% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogr-functions.R
\name{build_cod_matrix}
\alias{build_cod_matrix}
\title{Transform the COD data from a long table to a matrix with ages as rows and 
cod's as columns. Also replace na's with a very small number to avoid errors.}
\usage{
build_cod_matrix(y, vsn = 0)
}
\arguments{
\item{y}{COD long table}

\item{vsn}{very small scalar to replace na values}
}
\description{
Transform the COD data from a long table to a matrix with ages as rows and 
cod's as columns. Also replace na's with a very small number to avoid errors.
}
\keyword{internal}
