% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur_fit_obj.R
\name{.DollarNames.lemur_fit_obj}
\alias{.DollarNames.lemur_fit_obj}
\alias{$,lemur_fit_obj-method}
\alias{dollar_methods}
\alias{$<-,lemur_fit_obj-method}
\title{Access values from a \code{lemur_fit_obj}}
\usage{
\method{.DollarNames}{lemur_fit_obj}(x, pattern = "")

\S4method{$}{lemur_fit_obj}(x, name)

\S4method{$}{lemur_fit_obj}(x, name) <- value
}
\arguments{
\item{x}{the \code{lemur_fit_obj}}

\item{pattern}{the pattern from looking up potential values interactively}

\item{name}{the name of the value behind the dollar}

\item{value}{the replacement value. This only works for \code{colData} and
\code{rowData}.}
}
\description{
Access values from a \code{lemur_fit_obj}
}
\seealso{
\code{\linkS4class{lemur_fit_obj}} for more documentation on the
accessor functions.
}
