% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{chat}
\alias{chat}
\title{Helper to create chat objects}
\usage{
chat(role = character(), content = character())
}
\arguments{
\item{role}{A character vector of roles in the conversation ("system", "user" or "assistant")}

\item{content}{A character vector of messages in the conversation}
}
\description{
Helper to create chat objects
}
