% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur-data.R
\docType{data}
\name{data_gbd2021_lt}
\alias{data_gbd2021_lt}
\title{Abridged life table between 1990 and 2021 -- Global Burden of Disease Study 2022
Life tables constructed using the GBD2021 probability of death data.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 148500 rows and 13 columns.
}
\source{
Global Burden of Disease Collaborative Network.
Seattle, United States: Institute for Health Metrics and Evaluation (IHME), 2022.
Available from https://vizhub.healthdata.org/gbd-results/.
\href{https://vizhub.healthdata.org/gbd-results/}{
 Global Burden of Disease Study 2021 (GBD 2021) Results}
}
\usage{
data_gbd2021_lt
}
\description{
Abridged life table between 1990 and 2021 -- Global Burden of Disease Study 2022
Life tables constructed using the GBD2021 probability of death data.
}
\examples{
data_gbd2021_lt
}
\keyword{datasets}
