% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur-data.R
\docType{data}
\name{data_gbd2021_sdg}
\alias{data_gbd2021_sdg}
\title{Causes of Death Data between 1990 and 2021 -- Global Burden of Disease Study 2022}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3118500 rows and 6 columns.
}
\source{
Global Burden of Disease Collaborative Network.
Seattle, United States: Institute for Health Metrics and Evaluation (IHME), 2022.
Available from https://vizhub.healthdata.org/gbd-results/.
\href{https://vizhub.healthdata.org/gbd-results/}{
 Global Burden of Disease Study 2021 (GBD 2021) Results}
}
\usage{
data_gbd2021_sdg
}
\description{
In this data object the death count data (GBD2021) is grouped in such a way
that make possible the tracking of the evolution of the UN's
Sustainable Development Goals.
}
\examples{
data_gbd2021_sdg
}
\keyword{datasets}
