% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogr-functions.R
\name{decompose_by_age}
\alias{decompose_by_age}
\title{Perform decomposition of age-specific mortality contributions 
in life expectancy between any two populations}
\usage{
decompose_by_age(A, B)
}
\arguments{
\item{A}{Life table 1.}

\item{B}{Life table 2.}
}
\value{
A numerical vector.
}
\description{
Decompose the difference in life expectancy (at the smallest available age) 
of two populations represented by their life tables in a given 
period of time.
}
\examples{
# Data
X <- data_gbd2019_lt

# Select Life Table 1 & 2
L1 <- X[X$region == "Romania" & X$sex == "both" & X$level == "median", ]
L2 <- X[X$region == "Republic of Moldova" & X$sex == "both" & X$level == "median", ]

# Life expectancy at birth difference
L2$ex[1] - L1$ex[1] # about 2 years diff

# Age decomposition
decompose_by_age(L1, L2)
}
\references{
...
}
\author{
...
}
