% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{dt_filter_sql}
\alias{dt_filter_sql}
\title{Query data from a PostgresSQL. This would replace the local data and  the 
dt_filter() function}
\usage{
dt_filter_sql(data, mode, region1, region2, gender, year)
}
\description{
Query data from a PostgresSQL. This would replace the local data and  the 
dt_filter() function
}
\keyword{internal}
