% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-decompose.R
\name{exFUN}
\alias{exFUN}
\title{Life expectancy function}
\usage{
exFUN(x, cod)
}
\arguments{
\item{x}{vector of ages}

\item{cod}{Matrix containing cause-specific death rates}
}
\value{
life expectancy at birth (or min age)
}
\description{
Life expectancy function
}
\keyword{internal}
