% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_de_neighborhoods.R
\name{find_de_neighborhoods}
\alias{find_de_neighborhoods}
\title{Find differential expression neighborhoods}
\usage{
find_de_neighborhoods(
  fit,
  independent_matrix = NULL,
  group_by = NULL,
  selection_procedure = c("zscore", "contrast"),
  directions = c("random", "contrast", "axis_parallel"),
  independent_matrix_type = c("counts", "continuous"),
  de_mat = assay(fit, "DE"),
  contrast = fit$contrast,
  n_random_directions = 50,
  design = fit$design,
  include_complement = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fit}{the \code{lemur_fit} generated by \code{lemur()}}

\item{independent_matrix}{a matrix with the same dimensions as fit that contains
independent values and is used to confirm the significance of the neighborhoods.
The argument is optional, if it is provided the function returns additional
columns with the significance of each neighborhood based on a pseudobulked
differential expression test using \href{10.18129/B9.bioc.glmGamPoi}{\code{glmGamPoi}} or
\href{10.18129/B9.bioc.limma}{\code{limma}} (depending on \code{independent_matrix_type}).}

\item{group_by}{If the \code{independent_matrix} is provided, \code{group_by} defines
how the pseudobulks are formed.}

\item{selection_procedure}{specify the algorithm that is used to select the
neighborhoods for each gene. Broadly, \code{selection_procedure = "zscore"} is faster
but less precise than \code{selection_procedure = "contrast"}.}

\item{directions}{a string to define the algorithm to select the direction onto
which the cells are projected before searching for the neighborhood.
\code{directions = "random"} produces denser neighborhoods, whereas \code{directions = "contrast"}
has usually more power. \cr
Alternatively, this can also be a matrix with one direction for each gene
(i.e., a matrix of size \code{nrow(fit) * fit$n_embedding}).}

\item{independent_matrix_type}{is the \code{independent_matrix} argument a matrix of counts or
continuous values?}

\item{de_mat}{the matrix with the differential expression values and is only relevant if
\code{selection_procedure = "zscore"} or \code{directions = "random"}. Defaults
to an assay called \code{"DE"} that is produced by \code{lemur::test_de()}.}

\item{contrast}{a specification which contrast to fit. This defaults to the
\code{contrast} argument that was used for \code{test_de} and is stored in \code{fit$contrast}.}

\item{n_random_directions}{the number of random projections to use if \code{directions = "random"}.
Default: \code{50}.}

\item{design}{the design to use for the fit. Default: \code{fit$design}}

\item{include_complement}{a boolean to specify if the complement of the identified per gene
neighborhood is also returned. It will be marked in the output by \code{selection = FALSE}.}

\item{verbose}{Should the method print information during the fitting. Default: \code{TRUE}.}

\item{...}{additional parameters passed to underlying functions.}
}
\value{
a data frame with one entry per gene / neighborhood containing the name
of the neighborhood, the cell indices included in the neighborhood, the number of
cells, and the selection statistic. If \code{independent_matrix} is not \code{NULL}, the data frame will
also contain columns from the \code{limma} / \code{glmGamPoi} pseudobulk test (pval, adj_pval,
t_statistic / f_statistic, and lfc).
}
\description{
Find differential expression neighborhoods
}
