% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glioblastoma_example_data.R
\name{glioblastoma_example_data}
\alias{glioblastoma_example_data}
\title{The \code{glioblastoma_example_data} dataset}
\value{
A \code{\link{SingleCellExperimenbt}} object.
}
\description{
The dataset is a \code{\link{SingleCellExperiment}} object subset to 5,000 cells and
300 genes. The \code{colData} contain an entry for each cell from which patient
it came and to which treatment condition it belonged (\code{"ctrl"} or \code{"panobinostat"}).
}
\details{
The original data was collected by Zhao et al. (2021).
}
\references{
\itemize{
\item Zhao, Wenting, Athanassios Dovas, Eleonora Francesca Spinazzi, Hanna Mendes Levitin, Matei Alexandru Banu, Pavan Upadhyayula, Tejaswi Sudhakar, et al.
“Deconvolution of Cell Type-Specific Drug Responses in Human Tumor Tissue with Single-Cell RNA-Seq.” Genome Medicine 13, no. 1
(December 2021): 82. https://doi.org/10.1186/s13073-021-00894-y.
}
}
