% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_regression.R
\name{grassmann_geodesic_regression}
\alias{grassmann_geodesic_regression}
\title{Solve d(P, exp_p(V * x))^2 for V}
\usage{
grassmann_geodesic_regression(
  coordsystems,
  design,
  base_point,
  weights = rep(1, length(coordsystems)),
  tangent_regression = FALSE
)
}
\description{
Solve d(P, exp_p(V * x))^2 for V
}
