% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony_wrapper.R
\name{harmony_new_object}
\alias{harmony_new_object}
\title{Create an arbitrary Harmony object so that I can modify it later}
\usage{
harmony_new_object()
}
\description{
Create an arbitrary Harmony object so that I can modify it later
}
\keyword{internal}
