% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaiassistant-service.R
\name{initialise.openaiassistant}
\alias{initialise.openaiassistant}
\title{Initialise an openaiassistant chat}
\usage{
\method{initialise}{openaiassistant}(chat, json = FALSE)
}
\arguments{
\item{chat}{The chat object}

\item{json}{If TRUE, the model will respond in JSON. Defaults to FALSE}
}
\description{
Initialise an openaiassistant chat
}
