% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaiassistant-service.R
\name{initialise.openaiassistant}
\alias{initialise.openaiassistant}
\title{Initialise an openaiassistant chat}
\usage{
\method{initialise}{openaiassistant}(chat, json_schema = NA)
}
\arguments{
\item{chat}{The chat object}

\item{json_schema}{(Optional) A JSON schema that the response must strictly
adhere to}
}
\description{
Initialise an openaiassistant chat
}
