% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaiassistant-service.R
\name{initialise.openaiassistant}
\alias{initialise.openaiassistant}
\title{Initialise a chat}
\usage{
\method{initialise}{openaiassistant}(chat)
}
\arguments{
\item{chat}{The chat object}
}
\description{
Initialise a chat
}
