% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_embedding.R
\name{lemur}
\alias{lemur}
\title{Main function to fit the latent embedding multivariate regression (LEMUR) model}
\usage{
lemur(
  data,
  design = ~1,
  col_data = NULL,
  n_ambient = 30,
  n_embedding = 15,
  alignment = FALSE,
  base_point = c("global_embedding", "mean"),
  use_assay = "logcounts",
  ...,
  verbose = TRUE
)
}
\arguments{
\item{data}{a matrix with obsevations in the columns and features in the rows.
Or a \code{SummarizedExperiment} / \code{SingleCellExperiment} object}

\item{design}{a formula referring to global objects or column in the \code{colData} of \code{data}
and \code{col_data} argument}

\item{col_data}{an optional data frame with \code{ncol(data)} rows.}

\item{n_ambient}{the dimension of the ambient PCA. This should be large enough
to capture all relevant variation in the data}

\item{n_embedding}{the dimension of the $k$-plane that is rotated through space.
Needs to be smaller than \code{n_ambient}.}

\item{alignment}{optional specification how or if points should be aligned. This
can also be done in a separate step by calling \code{\link{align_embeddings}}.}

\item{base_point}{a string specifying how to find the base point for the geodesic
regression. Alternatively, an orthogonal matrix with \code{n_ambient} \eqn{\times} \code{n_embedding}
dimension.}

\item{use_assay}{if \code{data} is a \code{SummarizedExperiment} / \code{SingleCellExperiment} object,
which assay should be used.}

\item{...}{additional parameters that are passed on to the internal function \code{lemur_impl}.
For example, if you have pre-calculated the ambient PCA, you can provide it as a list with
the \code{coordsystem}, \code{embedding} and \code{offset} fields, to speed-up the fit.}

\item{verbose}{Should the method print information during the fitting. Default: \code{TRUE}.}
}
\description{
Main function to fit the latent embedding multivariate regression (LEMUR) model
}
