% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemur.R
\name{lemur}
\alias{lemur}
\title{Main function to fit the latent embedding multivariate regression (LEMUR) model}
\usage{
lemur(
  data,
  design = ~1,
  col_data = NULL,
  n_ambient = Inf,
  n_embedding = 15,
  alignment = FALSE,
  base_point = c("global_embedding", "mean"),
  use_assay = "logcounts",
  ...,
  verbose = TRUE
)
}
\arguments{
\item{data}{a matrix with observations in the columns and features in the rows.
Or a \code{SummarizedExperiment} / \code{SingleCellExperiment} object}

\item{design}{a formula referring to global objects or column in the \code{colData} of \code{data}
and \code{col_data} argument}

\item{col_data}{an optional data frame with \code{ncol(data)} rows.}

\item{n_ambient}{the dimension of the ambient PCA. This should be large enough
to capture all relevant variation in the data}

\item{n_embedding}{the dimension of the $k$-plane that is rotated through space.
Needs to be smaller than \code{n_ambient}.}

\item{alignment}{optional specification how or if points should be aligned. This
can also be done in a separate step by calling \code{\link{align_harmony}} or \code{\link{align_by_grouping}}.}

\item{base_point}{a string specifying how to find the base point for the geodesic
regression. Alternatively, an orthogonal matrix with \code{n_ambient} \eqn{\times} \code{n_embedding}
dimension.}

\item{use_assay}{if \code{data} is a \code{SummarizedExperiment} / \code{SingleCellExperiment} object,
which assay should be used.}

\item{...}{additional parameters that are passed on to the internal function \code{lemur_impl}.
For example, if you have pre-calculated the ambient PCA, you can provide it as a list with
the \code{coordsystem}, \code{embedding} and \code{offset} fields, to speed-up the fit.}

\item{verbose}{Should the method print information during the fitting. Default: \code{TRUE}.}
}
\value{
an object of class \code{lemur_fit} which extends \code{\link{SingleCellExperiment}}. Accordingly,
all functions that work for \code{sce}'s also work for \code{lemur_fit}'s. In addition, we
give easy access to the fitted values using the dollar notation (e.g., \code{fit$embedding}).
For details see the \linkS4class{lemur_fit} help page.
}
\description{
Main function to fit the latent embedding multivariate regression (LEMUR) model
}
\references{
\itemize{
\item Ahlmann-Eltze, C. & Huber, W. (2023). Analysis of multi-condition single-cell data with latent
embedding multivariate regression. bioRxiv \url{https://doi.org/10.1101/2023.03.06.531268}
}
}
\seealso{
\code{\link{align_by_grouping}}, \code{\link{align_harmony}}, \code{\link{test_de}}, \code{\link{find_de_neighborhoods}}
}
