% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\docType{package}
\name{LEMUR}
\alias{LEMUR}
\title{Linear mixed model-based differential Expression analysis accounting for Multiple samples Using Random effects}
\description{
Single cell differential expression analysis for multiple samples
}
\author{
Chih-Hsuan Wu \email{chihhsuan@uchicago.edu}
}
