% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-files.R
\name{list_files}
\alias{list_files}
\title{List files belonging to the user's organisation}
\usage{
list_files()
}
\value{
A data frame of files
}
\description{
List files belonging to the user's organisation
}
