% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-class.R
\name{messages}
\alias{messages}
\alias{messages.chat}
\title{Get messages from a chat}
\usage{
messages(chat, ...)

\method{messages}{chat}(chat, ...)
}
\arguments{
\item{chat}{The chat object}

\item{...}{Other service-specific arguments}
}
\value{
A data frame
}
\description{
Get messages from a chat
}
