% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-modify_LT.R
\name{modify_life_table}
\alias{modify_life_table}
\title{Modify life table by changing the cause of death associated risks}
\usage{
modify_life_table(lt, cod, cod_change)
}
\arguments{
\item{lt}{Life table;}

\item{cod}{Causes of death matrix containing death counts corresponding
to the population and time period of the life table;}

\item{cod_change}{Numerical scalar, vector or matrix.
The changes to be applied to the causes of death rate, m[c,x],
in order to reduce or increase the mortality estimates given by the life
table. Accepted input: any value greater than -100. See examples.}
}
\value{
A life table in the same format as the input life table.
}
\description{
Modify life table by changing the cause of death associated risks
}
