% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityCauses-package.R
\docType{package}
\name{MortalityCauses}
\alias{MortalityCauses}
\alias{MortalityCauses-package}
\title{MortalityCauses: Caused Deleted Mortality Shiny App}
\description{
What the package does (one paragraph).
}
\details{
To learn more about the package, start with the vignettes:
\code{browseVignettes(package = "MortalityCauses")}
}
\author{
\strong{Maintainer}: Marius D. Pascariu \email{rpascariu@outlook.com} (\href{https://orcid.org/0000-0002-2568-6489}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jose Manuel Aburto (\href{https://orcid.org/0000-0002-2926-6879}{ORCID})
  \item Vladimir Canudas-Romo (\href{https://orcid.org/0000-0001-6532-0089}{ORCID})
}

}
