% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-chat.R
\name{new_chat}
\alias{new_chat}
\title{A constructor function for chat}
\usage{
new_chat(x = list(role = character(), content = character()))
}
\arguments{
\item{x}{A list contatining a character vector of roles and a character vector of messages}
}
\description{
A constructor function for chat
}
