% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-figures.R
\name{plot_change}
\alias{plot_change}
\title{Plot the difference in life expectancy of two life tables}
\usage{
plot_change(
  L1,
  L2,
  age = seq(0, 110, by = 10),
  perc = FALSE,
  title = NULL,
  subtitle = ""
)
}
\arguments{
\item{L1}{Life table corresponding to the region/time period 1;}

\item{L2}{Life table corresponding to the region/time period 2;}

\item{age}{Reference ages.}

\item{perc}{Logical. If TRUE data will be displayed as percentages else
as absolute values. Default: FALSE.}

\item{title}{The text for the title.}

\item{subtitle}{The text for the subtitle for the plot which will be
displayed below the title.}
}
\description{
Plot the difference in life expectancy of two life tables
}
