% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-figures.R
\name{plot_cod}
\alias{plot_cod}
\title{Plot function for COD data}
\usage{
plot_cod(cod, perc = FALSE, type = "barplot")
}
\arguments{
\item{cod}{Causes of death matrix containing death counts corresponding
to the population and time period of the life table;}

\item{perc}{Logical. If TRUE data will be displayed as percentages else
as absolute values. Default: FALSE.}

\item{type}{Options: "barplot" or "piechart".}
}
\description{
Plot function for COD data
}
