% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-decompose.R
\name{plot_decompose}
\alias{plot_decompose}
\title{Plot Function for decompose}
\usage{
plot_decompose(object, ...)
}
\arguments{
\item{object}{An object of class decompose}

\item{...}{Arguments to be passed to methods, such as graphical}
}
\description{
Plot Function for decompose
}
\examples{
# See example in the ?decompose_by_cod or ?decompose_by_age help pages
}
\seealso{
\code{\link{decompose_by_cod}}
\code{\link{decompose_by_age}}
}
