% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-figures.R
\name{plot_map}
\alias{plot_map}
\title{Plot map}
\usage{
plot_map(location, zoom = 5, data = data_sf)
}
\arguments{
\item{location}{Geographical location}

\item{zoom}{The zoom level}

\item{data}{data}
}
\description{
Plot map
}
\examples{
plot_map(location = "MEXICO")

}
