% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_process_data.R
\name{prepare_data_mode_cod}
\alias{prepare_data_mode_cod}
\title{Prepare data for risk changes}
\usage{
prepare_data_mode_cod(cod, lt, region1, region2, cod_target, cod_change)
}
\description{
Prepare data for risk changes
}
\keyword{internal}
