% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{prepare_data_mode_sex}
\alias{prepare_data_mode_sex}
\title{Prepare data for sex comparisons}
\usage{
prepare_data_mode_sex(region1, cod_target, cod_change, year)
}
\description{
Prepare data for sex comparisons
}
\keyword{internal}
