% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant-class.R
\name{print.assistant}
\alias{print.assistant}
\title{A print method for the assistant class}
\usage{
\method{print}{assistant}(x, ...)
}
\arguments{
\item{x}{An assistant object}

\item{...}{Other arguments to be passed to print()}
}
\description{
A print method for the assistant class
}
