% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaiassistant-service.R
\name{print.openaiassistant}
\alias{print.openaiassistant}
\title{Print information about an openaiassistant chat}
\usage{
\method{print}{openaiassistant}(chat)
}
\arguments{
\item{chat}{The chat}
}
\description{
Print information about an openaiassistant chat
}
