% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaichat-service.R
\name{print.openaichat}
\alias{print.openaichat}
\title{Print information about an openaichat chat}
\usage{
\method{print}{openaichat}(x, ...)
}
\arguments{
\item{x}{The chat}

\item{...}{Further arguments passed from other methods}
}
\description{
Print information about an openaichat chat
}
