% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pseudoinverse}
\alias{pseudoinverse}
\title{Moore-Penrose pseudoinverse calculated via SVD}
\usage{
pseudoinverse(X)
}
\arguments{
\item{X}{a matrix X}
}
\description{
Moore-Penrose pseudoinverse calculated via SVD
}
\keyword{internal}
