% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_least_squares.R
\name{recursive_least_squares}
\alias{recursive_least_squares}
\alias{bulked_recursive_least_squares_contrast}
\title{Iteratively calculate the least squares solution}
\usage{
recursive_least_squares(y, X)

bulked_recursive_least_squares_contrast(y, X, group, contrast)
}
\arguments{
\item{y}{a vector with observations}

\item{X}{a design matrix}
}
\value{
a matrix where column i is the
solution to \code{y[1:i] ~ X[1:i,]}.
}
\description{
Iteratively calculate the least squares solution

Enable pseudobulking and directly calculate the contrast
}
\keyword{internal}
