% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{residuals,lemur_fit-method}
\alias{residuals,lemur_fit-method}
\title{Predict values from \code{lemur_fit} object}
\usage{
\S4method{residuals}{lemur_fit}(object, with_linear_model = TRUE, with_embedding = TRUE, ...)
}
\arguments{
\item{object}{an \code{lemur_fit} object}

\item{with_linear_model}{a boolean to indicate if the linear regression
offset is included in the prediction.}

\item{with_embedding}{a boolean to indicate if the embedding contributes
to the output.}

\item{...}{ignored.}
}
\description{
Predict values from \code{lemur_fit} object
}
