% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-class.R
\name{say}
\alias{say}
\alias{say.chat}
\title{Sent a message in a chat}
\usage{
say(chat, content, ...)

\method{say}{chat}(chat, ...)
}
\arguments{
\item{chat}{The chat object}

\item{content}{The message content}

\item{...}{Other service-specific arguments}
}
\description{
Sent a message in a chat
}
