% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-utils.R
\name{setSliderColor_}
\alias{setSliderColor_}
\title{shinyWidgets::setSliderColor (vendored)}
\usage{
setSliderColor_(color, sliderId)
}
\description{
Edit the color of Shiny's slider inputs, returning the CSS that recolors the
slider skin. Used internally to colour the app's sliders; not part of the
public API.
}
\keyword{internal}
