% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_regression.R
\name{spd_lm}
\alias{spd_lm}
\title{Solve ||Y - exp_p(V * x) Z ||^2_2 for V (aka. SPD Procrustes regression)}
\usage{
spd_lm(
  data,
  design,
  obs_embedding,
  base_point,
  ridge_penalty = 0,
  weights = NULL,
  tangent_regression = FALSE
)
}
\description{
Here data = t(grassmann_map(V * X)) Y
}
