% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stack_slice}
\alias{stack_slice}
\alias{destack_slice}
\title{Make a cube from a list of matrices}
\usage{
stack_slice(x)

destack_slice(x)
}
\arguments{
\item{x}{a list of vectors/matrices that are stacked}
}
\value{
A three-dimensional array.
}
\description{
The length of the list will become the third dimension of the cube.
}
\section{Functions}{
\itemize{
\item \code{destack_slice()}: Make a list of matrices from a cube

}}
\keyword{internal}
