% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API-files.R
\name{upload_file}
\alias{upload_file}
\title{Upload a file}
\usage{
upload_file(file, purpose = "assistants")
}
\arguments{
\item{file}{The file to upload}

\item{purpose}{The purpose of the file, usually either 'assistants' or 'fine-tuning'}
}
\description{
Upload a file
}
