% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant-class.R
\name{validate_assistant}
\alias{validate_assistant}
\title{A validator function for the assistant class}
\usage{
validate_assistant(x)
}
\arguments{
\item{x}{A assistant object}
}
\description{
A validator function for the assistant class
}
